% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory.R
\name{accessory}
\alias{accessory}
\alias{my.head}
\alias{my.tail}
\alias{rm.tail}
\alias{my.head.df}
\title{Accessory, internal functions}
\usage{
my.head(x, n = 1)

my.tail(x, n = 0)

rm.tail(x, n = 0)

my.head.df(x, n = 5)
}
\arguments{
\item{x}{a vector.}

\item{n}{a index.}
}
\description{
These accessory functions work similarly to \code{R} base functions
\code{head()} and \code{tail()} to return the first or last part of a vector or of a
data frame.
}
\details{
The functions \code{my.head()}, \code{my.tail()} and \code{rm.tail()} works with
vectors only, while function \code{my.head.df()} works with data frames or
matrices.
}
\examples{
my.head(letters)
my.tail(letters)
rm.tail(letters)
my.head.df(iris)
}
\keyword{internal}
