% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_tax.R
\name{accessory_tax}
\alias{accessory_tax}
\alias{rmOpen}
\alias{rmInfra}
\alias{rmHyb}
\alias{addRank}
\title{Accessory, internal functions for taxonomic manipulation}
\usage{
rmOpen(x)

rmInfra(x)

rmHyb(x)

addRank(x, rank = NULL)
}
\arguments{
\item{x}{a vector with scientific names to be standardized.}

\item{rank}{the expression or symbol to be added between names.}
}
\description{
These accessory functions work for editing the notation of
scientific names and are mainly used within \strong{plantR} function
\code{fixSpecies()}.
}
\details{
The functions \code{rmOpen()}, \code{rmInfra()}, and \code{rmHyb()} require only a
vector of scientific names, while the function \code{addRank()} also requires a
\code{rank} to be provided.

The function \code{rmOpen()} removes the open nomenclature 'cf.' and 'aff.'.

The function \code{rmInfra()} removes the infra-specific ranks from
varieties, sub-species and forms (e.g. 'var.', 'subsp.', 'f.')

The function \code{rmHyb()} removes the hybrid symbol 'x'.

The function \code{addRank()} does the opposite opperation: it adds ranks, opena
nomenclature or hybrid symbols into scientific names. For this function if
the number of ranks is equal to the number of names provided each rank is
assigned to the corresponding name. If the number of ranks is different,
the function silently uses the first rank for all names.
}
\examples{
\dontrun{
rmOpen(c("Lindsaea cf. lancea","Lindsaea aff. lancea"))
}

\dontrun{
rmInfra(c("Lindsaea lancea var. angulata",
          "Lindsaea lancea (L.) Bedd. var. angulata Rosenst.",
          "Lindsaea schomburgkii f. coriifolia (Lindm.) K.U. Kramer"))
}

\dontrun{
rmHyb(c("Blechnum ×antillanum", "Blechnum × antillanum",
       "Blechnum x antillanum", "Blechnum X antillanum"))
}

\dontrun{
addRank(c("Lindsaea lancea", "Lindsaea lancea angulata",
          "Lindsaea schomburgkii coriifolia", "Blechnum antillanum"),
          rank = c("cf.", "var.", "f.", "\u00d7"))
}


}
\seealso{
\link[plantR]{fixCase}.
}
\author{
Sara Mortara & Renato A. F. de Lima
}
\keyword{internal}
