% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_geo.R
\name{arw1}
\alias{arw1}
\title{Automatic Detection of Spatial Outliers}
\usage{
arw1(x, m0, c0, alpha = 0.025, pcrit)
}
\arguments{
\item{x}{a data frame with longitude and latitude in decimal degrees}

\item{m0}{center of the distribution}

\item{c0}{covariance matrix (p x p) of the distribution}

\item{alpha}{Maximum thresholding proportion. Default to 0.025.}

\item{pcrit}{Critical value obtained by simulations}
}
\description{
This is a local copy the function \code{mvoutlier::arw()}, that
implement adaptive reweighted estimator for location and scatter with
hard-rejection weights. It is only used in one specific function of
\strong{plantR}: \code{distOutlier()}.
}
\keyword{internal}
