% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bfoNames}
\alias{bfoNames}
\title{Brazilian Flora Online Vascular Plant Taxonomy}
\format{
An object of class \code{data.frame} with 14 columns and
over 125 thousand rows.
}
\source{
\url{https://ipt.jbrj.gov.br/jbrj/resource?r=lista_especies_flora_brasil}
}
\usage{
data(bfoNames)
}
\description{
A dataset containing the essential taxonomic
information of vascular plants stored in the
\href{https://floradobrasil.jbrj.gov.br/consulta/}{Brazilian Flora 2020}
taxonomic backbone, a.k.a. the Flora and Funga of Brazil. It
includes all taxonomic levels (i.e. infra-species, species,
genus, family and so on). See package \strong{plantRdata} for more
details on how this table was obtained.
}
\keyword{datasets}
\note{Last update/change of the downloaded backbone (day month year):
\code{01 Nov 2025}.}
