% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_tax.R
\name{buildName}
\alias{buildName}
\title{Build Organism Name}
\usage{
buildName(x, col.names = c("genus", "specificEpithet"))
}
\arguments{
\item{x}{the data frame with the name information}

\item{col.names}{the name of the columns containing the information to be
combined (in the desired order)}
}
\value{
a vector with the combined name information
}
\description{
Combine diffent table columns with name information (i.e.
genus, epiteth, infra-epiteth) into a single name
}
\author{
Renato A. F. de Lima
}
\keyword{internal}
