% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchAll.R
\name{catchAll}
\alias{catchAll}
\title{Catch Values, Warnings and Errors}
\usage{
catchAll(fun)
}
\arguments{
\item{fun}{a function.}
}
\value{
a list with the output of the function in the first element,
the warnings in the second and the errors in the third.
}
\description{
The function separates the output from another function, its
warning and errors, without printing them in the console.
}
\details{
The function was taken from the following stackoverflow question:
https://stackoverflow.com/questions/4948361/how-do-i-save-warnings-and-errors-as-output-from-a-function/4952908#4952908
}
\keyword{internal}
