% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCoord.R
\name{checkCoord}
\alias{checkCoord}
\title{Check Geographical Coordinates}
\usage{
checkCoord(x = occs, lon = "decimalLongitude.new", lat = "decimalLatitude.new")
}
\arguments{
\item{x}{Data.frame with coordinates in decimal degrees.}

\item{lon}{Column with the longitude to be validated}

\item{lat}{Column with the latitude to be validated}
}
\description{
This function makes the first check of the coordinates
against the worldMap and LatamMap shapefiles
}
\author{
Andrea Sánchez-Tapia, Sara Mortara & Renato A. F. de Lima
}
