% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInverted.R
\name{checkInverted}
\alias{checkInverted}
\title{Checks records for inverted and transposed coordinates}
\usage{
checkInverted(
  x,
  country_gazetteer = "country",
  lat = "decimalLatitude.new",
  lon = "decimalLongitude.new"
)
}
\arguments{
\item{x}{Occurrences data frame}

\item{country_gazetteer}{Column with the country according to the exsicatta}

\item{lat}{Column with the corrected clean latitude Defaults to decimalLatitude.new}

\item{lon}{Column with the corrected clean latitude Defaults to decimalLongitude.new}
}
\description{
Checks records for inverted and transposed coordinates
}
\author{
Andrea Sánchez-Tapia & Sara Mortara
}
