% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklist.R
\name{checklist}
\alias{checklist}
\title{Create Species Checklists}
\usage{
checklist(
  x,
  fam.order = TRUE,
  n.vouch = 30,
  type = "short",
  rm.dup = TRUE,
  type.rank = 5,
  date.format = "\%d \%b \%Y"
)
}
\arguments{
\item{x}{a data frame with the occurrence data, generally as the output of the
__plantR__ validation functions.}

\item{fam.order}{logical. Should taxa be listed under their families?
Default to TRUE.}

\item{n.vouch}{numerical. Maximum number of vouchers to be listed. Default
to 30.}

\item{type}{character. The type of voucher list desired. Options are: 'short',
'selected' and 'list' (see details below).}

\item{rm.dup}{logical. Should duplicated specimens be removed prior to the
calculation of species summaries? Default to TRUE.}

\item{type.rank}{numerical. Value to be assigned to the rank of type
specimens in order to organize and filter the coucher list. Default to 5.}

\item{date.format}{The desired format for the dates. Default to "\%d \%b \%Y".}
}
\description{
This function creates a checklist of the species contained
in the occurrence data.
}
\details{
The checklist can be organized by alphabetic order of the taxa or by
  alphabetic order of families (default).

  By default, the checklist provides the number of records found and the
  overall taxonomic and geographic confidence level of the records (columns
  'tax.CL' and 'geo.CL'). The taxonomic confidence level is the percentage of
  records with the identification flagged as 'high', while the geographic
  confidence level is the percentage of records with coordinates flagged as
  being validated at municipality or locality levels.

  The function also provide a list of vouchers, giving priority to type
  specimens and records with higher level of taxonomic confidence. By
  default, the function provides only up to 30 vouchers, but this number can
  be controlled using the argument `n.vouch`.

  The voucher list can be provided in the following output formats (The
  option 'list' is not implemented yet):
  \itemize{
  \item 'short': Collector name, Collector number (collections of deposit)
  \item 'selected': COUNTRY, stateProvince: municipality, Date, Collector name,
  Collector number (collections of deposit)
  \item 'list': Collector name, Collector number(s) (species code)
  }

  Note: although we provide a `date.format` argument, formats other then the
  default are pending and so they may not work properly.
}
