% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSea.R
\name{checkSea}
\alias{checkSea}
\title{Checks occurrences that fall in the sea and close to the shores}
\usage{
checkSea(
  x = data.frame(occs),
  lat = "decimalLatitude.new.new",
  lon = "decimalLongitude.new.new"
)
}
\arguments{
\item{x}{Dataframe with species occurrence}

\item{lat}{Column with the latitude to be checked. Defaults to *.new.new that was corrected for inverse coordinates}

\item{lon}{Column with the latitude to be checked. Defaults to *.new.new that was corrected for inverse coordinates}
}
\description{
Checks occurrences that fall in the sea and close to the shores
}
\author{
Andrea Sánchez-Tapia & Sara Mortara
}
