% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanName.R
\name{cleanName}
\alias{cleanName}
\title{Remove punctuation, special characters, numbers and spaces}
\usage{
cleanName(x)
}
\arguments{
\item{x}{a vector of characters (i.e. names)}
}
\value{
the vector \code{x} without punctuation, special
characters, numbers and spaces
}
\description{
Simple function to prepare vectors of taxon names for
increasing the chances of exact and fuzzy name matching
}
\examples{

\dontrun{
taxa <- c("Genus \u00d7 epiteth (Jos\u00e9 AB.) C.")
print(taxa)
cleanName(taxa)
}

}
\author{
Renato A. F. de Lima
}
\keyword{internal}
