% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colNumber.R
\name{colNumber}
\alias{colNumber}
\title{Format Collector Number}
\usage{
colNumber(x, colCodes = NULL, noNumb = "s.n.")
}
\arguments{
\item{x}{the character string.}
}
\value{
the character string \code{x}
}
\description{
Cleans the the collector number field from herbarium labels.
}
\details{
The function puts
}
\examples{
# A vector with some typical examples of formats found in herbarium labels
numbers = c("3467","3467 "," 3467","ALCB3467","Gentry 3467","ALCB-3467","3467a","3467A","3467 A",
"3467-A","PL671","57-685","685 - 4724","1-80","-4724","(3467)","(3467","3467)","32-3-77",
"s/n.","s.n.","s.n","s/nº","",NA)
# Using the function defaults
colNumber(numbers)
# Using the function to remove the collection code from the collector number
colNumber(numbers, colCodes = c("ALCB","ESA"))
# Defining user-specific abbreviations for the specimens without collector number
colNumber(numbers, colCodes = c("ALCB","ESA"), noNumb = "n.a.")

}
\references{
Willemse, L.P., van Welzen, P.C. & Mols, J.B. (2008). Standardisation in data-entry across databases: Avoiding Babylonian confusion. Taxon 57(2): 343-345.
}
\author{
Lima, R.A.F.
}
