% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_geo.R
\name{distOutlier}
\alias{distOutlier}
\title{Outlier Mahalanobis Distances}
\usage{
distOutlier(
  lon,
  lat,
  method = "robust",
  n.min = 10,
  digs = 3,
  probs = c(0.5, 0.75, 0.9, 0.95, 0.975, 0.99, 1)
)
}
\arguments{
\item{lon}{numerical. Longitude in decimal degrees}

\item{lat}{numerical. Latitude in decimal degrees}

\item{method}{character. Type of method desired: 'classic' or 'robust'}

\item{n.min}{numerical. Minimun number of unique coordinates to be used in
the calculations. Default to 10}

\item{digs}{numerical. Number of digits to be returned after the decimal point.
Default to 4}

\item{probs}{numerical. Vector of probabilities between 0 and 1 to calculate
the sample quantiles. Defaul to c(0.5, 0.75, 0.9, 0.95, 0.975, 0.99, 100)}
}
\value{
the number of unique coordinates ('n'), the number of outliers
detected ('n.out') and the sample quantiles ('qt') of the Mahalanobis
distances.
}
\description{
Outlier Mahalanobis Distances
}
\details{
The function returns the quantiles of the Mahalanobis distances for
the spatial outliers detected automatically, which can be used in the decision making of
the more appropriated distance cutoffs to flag spatial outliers.

The automatic detection of spatial outliers is based on an adjusted threshold
of the Mahalanobis distances using package \code{mvoutlier} and its defaults
(function \code{mvoutlier::arw()}).

If the number of unique coordinates is below \code{n.min} or if the Minimum
Covariance Determinant (MCD) estimator has issues, the function returns NAs.
See the help of function \code{mahalanobisDist()} for details on the other
parameters.
}
\examples{
set.seed(123)
lon <- runif(100, -45, -35)
lat <- runif(100, -35, -25)

distOutlier(lon, lat, method = "classic") # no outliers found...
distOutlier(lon, lat, method = "robust")

# adding some noise
lat[1:5] <- lat[1:5] + runif(5, 10, 20)
lon[96:100] <- lon[96:100] + runif(5, 10, 20)

distOutlier(lon, lat, method = "classic") # quantiles for the 10 outliers found
distOutlier(lon, lat, method = "robust")

}
\seealso{
\link[plantR]{mahalanobisDist}.
}
\author{
Renato A. Ferreira de Lima
}
