% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downFBO.R
\name{downFBO}
\alias{downFBO}
\title{Download Species Information from the Brazilian Flora Online}
\usage{
downFBO(
  id,
  campos = c("estadosCerteza", "descricaoCamposControlados", "descricaoLivrePT",
    "vouchers", "endemismo", "origem", "substrato", "dominioFitogeografico", "formaVida",
    "tipoVegetacao", "id", "nomeStr", "bibliografiaReferencia", "referencias")
)
}
\arguments{
\item{id}{Vector. A list of taxon IDs from the Brazilian Flora database.}

\item{campos}{Vector. A list of fields to be returned from the Brazilian
Flora database.}
}
\description{
The function makes an online query for informations stored in the
\href{http://floradobrasil.jbrj.gov.br/reflora/listaBrasil}{Brazilian Flora Online}.
}
\keyword{internal}
