% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example}
\alias{example}
\title{Occurrence data of 300 species/morphospecies downloaded from speciesLink blabla}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 504 rows and 35 columns.}
\usage{
example
}
\description{
Occurrence data of 300 species/morphospecies downloaded from speciesLink blabla
}
\keyword{datasets}
