% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_intro}
\alias{example_intro}
\title{Dataset for the package introductory vignette}
\format{
A data frame with 713 rows and 45 variables
}
\source{
\url{https://specieslink.net/}
}
\usage{
data(example_intro)
}
\description{
A dataset containing the retrieved records of the species
\emph{Euterpe edulis} downloaded from the speciesLink network
in May 2024. Data was filtered to contain only the columns most
relevant to \code{plantR}
}
\keyword{datasets}
