% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_pecb}
\alias{example_pecb}
\title{Dataset to illustrate the data editing and validation}
\format{
A data frame with 504 rows and 35 variables
}
\usage{
data(example_pecb)
}
\description{
A dataset containing the retrieved records downloaded from JABOT, which
has the typical fields and format used by different herbaria.
}
\keyword{datasets}
