% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixAnnotation.R
\name{fixAnnotation}
\alias{fixAnnotation}
\title{Fix Taxon Name Notation}
\usage{
fixAnnotation(x)
}
\arguments{
\item{x}{a vector of characters with taxon names}
}
\value{
the vector \code{x} with the standardize taxon name
notation
}
\description{
Standardizes the notation of open nomenclature
and modificators of taxon names (e.g. aff., cf., etc.).
}
\details{
The function solve most but not all possible notation of
name modificators, specially when there are spaces that separate
them from the naes themselves are missing (see examples).
}
\examples{
\dontrun{
taxa <- c("Lindsaea lancea var angulata", "Lindsaea Aff.lancea",
"Blechnum cf.spannagelii", "Blechnumcf.spannagelii")
fixAnnotation(taxa)
}

}
\author{
Renato A. F. de Lima
}
\keyword{internal}
