% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixAuthors.R
\name{fixAuthors}
\alias{fixAuthors}
\title{Fix Taxon Name with Authorship}
\usage{
fixAuthors(taxa = NULL, ranks = c("var.", "subsp.", "f.", "forma"))
}
\arguments{
\item{taxa}{a vector containing taxon names}

\item{ranks}{a vector containing the super or infra-specific ranks
to be searched and corrected within problematic taxon
name-authorship splits. Defaults to 'var.', 'subsp.', 'f.' and
'forma'}
}
\value{
a data frame with the original name, the cleaned scientific
and the authorship in different columns
}
\description{
Recognize strings of scientific names with authorship
and split scientific name and authorship into two different
columns of a data frame.
}
\details{
The function works for most cases but it is not perfect,
particularly if authorship notation in different of what is
commonly expected. The expectation are names following the
notation:
\itemize{
\item Higher-taxa authorship: e.g. 'Araceae Juss.'
\item Genus epithet authorship: e.g. 'Anthurium amoenum Kunth & C.D.Bouche'
\item Genus epithet infraepithet authorship: e.g. 'Anthurium amoenum var. humile (Schott) Engl.'
}

Besides rare examples of author names starting with lower-case
prepositions (e.g. 'de', 'van'), authorships should have
the first letter capitalized, otherwise the function will only
recognize the start of the authorship for very common author
names (e.g. Kuntze, Baker, Nees, Klotzsch, Mez). The full list of
common author names is stored in the internal \strong{plantR} object
\code{commonAuthors}.

In addition, the function is case-sensitive and it assumes that
there are no issues related to scientific name casing (e.g.
'Anthurium cf. Amoenum') to separate the scientific name from the
authorship. So, please make sure there are no casing issues. See
the \strong{plantR} function \code{fixCase()}
}
\examples{
names <- c("Lindsaea", "Lindsaea sp.",
 "Lindsaea lancea",
 "Lindsaea lancea (L.) Bedd.",
 "Parablechnum C.Presl",
 "Blechnum spannagelii Rosenst.",
 "Blechnum austrobrasilianum de la Sota",
 "Blechnum occidentale leopoldense Dutra",
 "Casearia sylvestris var. angustifolia",
 "Casearia sylvestris var. angustifolia Uittien",
 "Casearia sylvestris Sw. var. sylvestris",
 "Philodendron sodiroi hort.",
 "Richterago kuntze")

fixAuthors(names)

}
\seealso{
\link[plantR]{fixCase}
}
\author{
Renato A. Ferreira de Lima
}
