% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixEncoding.R
\name{fixEncoding}
\alias{fixEncoding}
\title{Fix Encoding Problems}
\usage{
fixEncoding(x, encoding = NULL)
}
\arguments{
\item{x}{a vector with characters to be replaced}

\item{encoding}{a character containing the encoding used to mark the output
vector}
}
\value{
the input vector with the encoding problems replaced
}
\description{
Simple function to replace typical encoding problems when
importing text with accents or special characters which is not in UTF-8
(i.e. latin1 encoding).
}
\details{
The function makes no encoding guessing. It just directly replaces
common enconding problems from text in latin1 to UTF-8 characters. So, make
sure that this replacement makes sure to your data.
}
\examples{
  nomes <- c("José", "Gonçalves", "Müller", "Londoño")

\dontrun{
  fixEncoding(nomes)
  fixEncoding(nomes, encoding = "UTF-8")
  fixEncoding(nomes, encoding = "latin1")

}

}
\author{
Renato A. F. de Lima
}
\keyword{internal}
