% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixIndet.R
\name{fixIndet}
\alias{fixIndet}
\title{Fix Taxon Name Notation}
\usage{
fixIndet(x, rplc = "Indet. sp.")
}
\arguments{
\item{x}{a vector of characters with taxon names}

\item{rplc}{a character with the standard to be used as the
replacement for the incomplete identification. Defaults to
"Indet. sp.".}
}
\value{
the vector \code{x} with the standardize notation of the
incomplete species identification
}
\description{
Standardizes the notation of taxon names with
incomplete identifications (i.e. not at the species level).
}
\details{
The function solve most but not all possible notation of
incomplete identifications, specially when the notation to
separate morphotypes include letters instead of numbers.
}
\examples{
taxa <- c("Indeterminada1", "Indeterminada sp1", "undeterminedA",
  "Fabaceae1", "FabaceaeA", "indet", "Fabales2", "Andira sp. 2",
  "Andira sp 2", "Andira sp")
fixIndet(taxa)

}
\author{
Renato A. Ferreira de Lima
}
