% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixInverted.R
\name{fixInverted}
\alias{fixInverted}
\title{Fixes inverted coordinates according to checkInverted}
\usage{
fixInverted(x = occs)
}
\arguments{
\item{x}{Data frame with species occurrences}
}
\description{
Fixes inverted coordinates according to checkInverted
}
\author{
Andrea Sánchez-Tapia & Sara Mortara
}
