% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixSpecies.R
\name{fixSpecies}
\alias{fixSpecies}
\title{Edit Scientific Name}
\usage{
fixSpecies(
  x = NULL,
  tax.name = "scientificName",
  rm.rank = FALSE,
  rm.indet = FALSE
)
}
\arguments{
\item{x}{a data.frame containing the species name}

\item{tax.name}{character. Name of the columns containing the species name.
Default to "scientificName"}

\item{rm.rank}{logical. Should the infra-specific rank abbreviation be
removed from the name? Default to FALSE}

\item{rm.indet}{logical. Should the abbreviations for unespecific names (i.e.
sp. or spp.) be removed? Default to FALSE}
}
\value{
Data frame with \code{verbatimSpecies} as in the original input, \code{speciesStatus}
with the flags in original data and \code{species} with a suggestion for a more
correct species name. See Details for a description of flags in the column
\code{speciesStatus}.
}
\description{
Identifies open nomenclature (aff., cf.) in species scientific name,
classification under species level (var. and subsp.), and common mistakes in
the species scientific name. It creates a new column with the original
string and the new suggested name. It also flags problematic names (character
string with numbers, authors, wrong case, or other names besides genus and
epithet etc). Names can be return with or without infra-specific ranks (var.
and subsp.) or abbreviations of unespecific names (sp. or spp.).
}
\details{
Possible flags returned in \code{species_status}: \describe{
\item{\code{possibly_ok}}{scientific name following the expected pattern 'Genus
epithet'}
\item{\code{not_Genus_epithet_format}}{scientific name not following
the expected pattern Genus epithet}
\item{\code{variety}}{species scientific name with variety}
\item{\code{subspecies}}{species scientific name with subspecies}
\item{\code{form}}{species scientific name with form}
\item{\code{confer}}{open nomenclature cf. in species scientific name}
\item{\code{affinis}}{open nomenclature aff. in species scientific name}
\item{\code{name_w_authors}}{species scientific name has authors}
\item{\code{not_name_has_digits}}{species scientific name has digits, not a
valid name}
\item{\code{indet}}{species identified only at genus level}
\item{\code{family_as_genus}}{family as genus, not a valid name}
\item{\code{order_as_genus}}{order as genus, not a valid name}
\item{\code{species_nova}}{species name contains an indication of a new
species, possibly not yet a valid name}
\item{\code{non_ascii}}{species name
has non ASCII characters, not a valid name}
\item{\code{hybrid_species}}{hybrid species} }
}
\examples{
df <- data.frame(scientificName =
c("Lindsaea lancea", "Lindsaea lancea (L.) Bedd.",
"Lindsaea lancea var. angulata",
"Lindsaea Aff. lancea",
"Lindsaea", "Lindsaea sp.", "Lindsaeaceae sp.",
"Lindsaea aff. lancea (L.) Bedd."))

fixSpecies(df)
fixSpecies(df, rm.rank = TRUE)
fixSpecies(df, rm.rank = TRUE, rm.indet = TRUE)

}
\references{
Sigovini, M., Keppel, E. and Tagliapietra, D. (2016) Open Nomenclature in the
biodiversity era. Methods in Ecology and Evolution 7(10): 1217-1225.
}
\author{
Sara Mortara
}
