% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatLoc.R
\name{formatLoc}
\alias{formatLoc}
\title{Format Locality Information}
\usage{
formatLoc(x, ...)
}
\arguments{
\item{x}{a data frame, containing typical fields from occurrence records from
herbarium specimens}
}
\value{
The input data frame \code{x}, plus the new columns with the formatted
  information.
}
\description{
This function standardizes names of administrative levels from
  species occurrences obtained from on-line databases, such as GBIF or
  speciesLink
}
\details{
The function works as a wrapper, where the individuals steps of the
  proposed __plantR__ workflow for editing locality information are performed
  altogether (see the __plantR__ tutorial for details).

  The input data frame usually contains the following locality fields:
  "country", "stateProvince", "municipality" and "locality".
}
\seealso{
\link[plantR]{fixLoc}, \link[plantR]{strLoc}, \link[plantR]{prepLoc},
 and \link[plantR]{getLoc}.
}
