% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatName.R
\name{formatName}
\alias{formatName}
\title{Format People's Name In TDWG Format}
\usage{
formatName(x)
}
\arguments{
\item{x}{the character string.}
}
\value{
the character string \code{x} in the fixed TDWG format. If
\code{x} is not in the TDWG format, the function returns \code{x}.
}
\description{
Fix minor problems on names already provided in the TDWG format.
}
\details{
The function fixes problems (e.g. missing points between name initials)
for names provided already in the TDWG format (i.e. last name + comma + initials).
}
\examples{
  # Simple name
  formatName("Gentry, AH")
  formatName("GENTRY, AH")
  formatName("Gentry, A H")
  formatName("Gentry, Alwyn Howard")

  # does nothing (not in TDWG format)
  formatName("Gentry AH")
  formatName("Gentry A.H.")

  # Names with generational suffixes
  formatName("Leitao Filho, HF")
  formatName("Filho Neto, S J")
  formatName("Leitao filho, H. F.")

  # Compound last name
  formatName("Saint-Hilaire, Augustin")
  formatName("Ziffer Berger, Richter")
  formatName("Saint-hilaire, A.")
  formatName("Saint-Hilaire A.")

  # Unusual formatting (function won't work always...)
  formatName("Sa, Cyl")
  # one name, two commas: fails to get all names
  formatName("Cesar Sandro, Esteves, F")
  # two or more names: output incorrect (combine names of authors)
  formatName("Mendonca Filho, C.V.; Neto, F.C.C.")
  # two or more names, separeted by comma: output incorrect (combine names of authors)
  formatName("A. Alvarez, A. Zamora & V. Huaraca")
}
\references{
Willemse, L.P., van Welzen, P.C. & Mols, J.B. (2008). Standardisation in data-entry across databases: Avoiding Babylonian confusion. Taxon 57(2): 343-345.
}
\author{
Lima, R.A.F.
}
