% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatSpecies.R
\name{formatSpecies}
\alias{formatSpecies}
\title{Get Valid Plant Names}
\usage{
formatSpecies(
  x,
  col.names = c("scientificName.new", "scientificNameAuthorship"),
  db = c("bfo", "tpl"),
  sug.dist = 0.9,
  use.authors = TRUE,
  drop.cols = c("ordem", "family", "verbatimSpecies", "author", "full_sp",
    "authorship", "id")
)
}
\arguments{
\item{x}{a data.frame containing the species scientific name without authors
and their authorship, ideally already passed by the string check in `fixSpecies`.
The authorship is only used for solving homonyms.}

\item{col.names}{character. Names of the columns containing the species names and authors.
Default to 'scientificName.new' and 'scientificNameAuthorship'.}

\item{db}{the list of database to be consulted for valid names, in their
preferred order of priority. Only the results from Brazilian Flora 2020
('bfo'), The Plant List ('tpl') or both are currently implemented.}

\item{sug.dist}{a fraction expressing the maximum distance allowed between
the original species name and the suggested species name, which is passed
to the arguments `suggestion.distance` of function `flora::get.taxa` and
`max.distance` of function `Taxonstand::TPL`. Default to 0.9.}

\item{use.authors}{logical. Should all authors names be verified (takes
longer)? Default to TRUE.}

\item{drop.cols}{character. Name of columns that should be dropped from the
results.}
}
\value{
A data frame with the name provided by the user and ...
}
\description{
Function that try to provide valid plant names or to correct
  typos and ortographical variants based on the
  \href{http://floradobrasil.jbrj.gov.br/reflora/listaBrasil/ConsultaPublicaUC/ConsultaPublicaUC.do}{Brazilian
   Flora 2020} project and on the \href{http://www.theplantlist.org/}{The
  Plant List}.
}
\examples{
sp.list <- c("Casearia sylvestris","Casearia sylvestirs",
  "Casearia sylvestris angustifolia", "Casearia attenuata",
  "Casearia celtidifolia", "Casearia celtidifolia","Casearia tropicana")
aut_list <- c("Sw.", "Sw.", "Uittien", "Rusby", "Kunth", "de Vriese","L.")
df <- data.frame(scientificName.new = sp.list,
  scientificNameAuthorship = aut_list,
  stringsAsFactors = FALSE)
formatSpecies(df)
formatSpecies(df, db = "bfo")
formatSpecies(df, db = "tpl")
formatSpecies(df, db = "tpl", sug.dist = 0.85)

}
\author{
Renato A. Ferreia de Lima
}
