% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTax.R
\name{formatTax}
\alias{formatTax}
\title{Format Taxonomical Nomenclature}
\usage{
formatTax(x)
}
\arguments{
\item{x}{data.frame}
}
\description{
This function standardize and validate the names of plant species and families.
}
