% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_geo.R
\name{geoDist}
\alias{geoDist}
\title{Distance between Coordinates}
\usage{
geoDist(lon, lat, radius = 6371)
}
\arguments{
\item{lon}{numerical. Longitude in decimal degrees}

\item{lat}{numerical. Latitude in decimal degrees}

\item{radius}{numerical. Radius in kilometers to find spherical distances.
Default to 6371 km. (If 1 distances are returned in radians)}
}
\value{
A distance matrix
}
\description{
Calculate the distances between geographical coordinates, taking
into account that Earth is an oblate spheroid. By default, distances are
returned in kilometers. This function is adapted from the function
\code{rdist.earth()} from package \strong{fields}.
}
\examples{

lon = c(-47, -46, -47)
lat = c(-23, -24, -23.5)

\dontrun{
geoDist(lon, lat)
}

}
\references{
Nychka, D., Furrer, R., Paige, J., Sain, S. (2017). fields: Tools for spatial
data. R package version 11.6, <URL: https://github.com/NCAR/Fields>.
}
\author{
Renato A. F. de Lima
}
\keyword{internal}
