% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdmin.R
\name{getAdmin}
\alias{getAdmin}
\title{Get Administrative Names}
\usage{
getAdmin(x, admin.names = "plantR")
}
\arguments{
\item{x}{a character or vector containing the locality search strings}

\item{admin.names}{a data frame containing the list of names of the locality subdivisions;
the default is "plantR" (biased towards Latin America)}
}
\value{
...
}
\description{
Obtain the names of the administrative subdivisions for a given locality from the gazetteer.
}
\details{
The function ...
}
\author{
Lima, R.A.F.
}
