% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxBackbone.R
\name{getTaxBackbone}
\alias{getTaxBackbone}
\title{Get Reference Taxonomy}
\usage{
getTaxBackbone(db)
}
\arguments{
\item{db}{a character with the taxonomic backbone acronym or a data
frame or list of data frames containing the taxonomic backbones.}
}
\value{
a list with as many elements as the length of db
}
\description{
Users can chose between the taxonomic backbone for
vascular plants of the Brazilian Flora and Funga (the internal
\code{plantR} dataset called \code{bfoNames} - the default) or to provide a
data frame (or a list of data frames) with the taxonomic
backbone(s) of their choice.
}
\details{
The expected format of the user-provided backbone is a data frame
or a list of data frame that must contain at least the following
columns: 'name' and 'authorship'. But ideally, it should have
more columns such as the backbone taxon ID (column 'id'), other
information on the taxa (i.e. columns 'phylum', 'family',
'taxon.rank', 'taxon.status', 'name.status') and on the possibly
accepted taxa (i.e. columns 'accepted.name', 'accepted.authorship',
'accepted.taxon.rank', 'accepted.taxon.status',
'accepted.name.status'). See the object 'bfoNames' for an example
of the expected format of the taxonomic backbone.
}
\examples{

dbs <- getTaxBackbone("bfo")


}
