% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxNotes.R
\name{getTaxNotes}
\alias{getTaxNotes}
\title{Get Taxonomic Notes}
\usage{
getTaxNotes(
  x = NULL,
  match.cols = c("suggestedName", "match_type", "multiple_match", "fuzzy_dist_name",
    "fuzzy_dist_author", "name.status", "taxon.status"),
  no.match = "no_match"
)
}
\arguments{
\item{x}{a data frame containing the results of the taxon name
matching against a reference backbone}

\item{match.cols}{a vector of column names containing the
information on the name matching process. Defaults to the
typical names used internally within the \code{plantR} function
\code{prepSpecies()}}

\item{no.match}{a character indicating the tag for taxon names
without match. Defaults to 'no_match'}
}
\value{
the same input data frame 'x' with an extra column called
'notes'
}
\description{
Creates more user-friendly categories regarding the
taxonomic and nomenclature status of taxon names based on the
output of the taxon name checking process.
}
\details{
The function returns the following categories (or
combination between them separated by a vertical bar): \describe{
\item{\code{accepted}}{A valid and accepted taxon name}
\item{\code{synonym}}{A valid name but that taxon that now goes by
a different accepted name}
\item{\code{name and/or author misspelled}}{A valid and accepted
taxon name that was written with typos on the taxon name and/or
authorship}
\item{\code{check +1 name}}{A name, generally without authorship
information, that has more than one accepted name}
\item{\code{check not resolved}}{A name currently not resolved that
has an 'unchecked', 'unplaced', 'doubtful' or empty taxonomic
status category in the reference backbone}
\item{\code{not found}}{A name not found after exact and fuzzy
matching, given the suggested maximum distance between input and
reference names }
\item{\code{bad match}}{A name found after fuzzy matching, but that
is finally above the suggested maximum distance (generally in the
case of large name + authorship combinations)} }
}
\author{
Renato A. Ferreira de Lima
}
