% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxUnique.R
\name{getTaxUnique}
\alias{getTaxUnique}
\title{Get Unique Taxon Names}
\usage{
getTaxUnique(
  df = NULL,
  df.ref = NULL,
  match.col = NULL,
  orig.col = "scientificName.new",
  name.col = "tax.name",
  status.col = "taxon.status",
  type.match.col = "match_type",
  mult.match.col = "multiple_match",
  mult.matches = "all",
  agg.cols = c("id", "family", "tax.authorship", "taxon.rank", "name.status",
    "taxon.status", "accepted.tax.name", "accepted.tax.authorship",
    "accepted.taxon.rank", "accepted.taxon.status", "accepted.name.status")
)
}
\arguments{
\item{df}{a data frame containing the input names}

\item{df.ref}{a data frame containing the reference names}

\item{match.col}{a vector with the name of the column in both
datasets to be used for matching}

\item{orig.col}{a vector with the name of the column with the
original name in the input data frame 'df'. Defaults to
"scientificName.new" which is the output of the \code{plantR} function
\code{fixSpecies()}}

\item{name.col}{a vector with the name of the column with the taxon
name in the data frame 'df.ref'. Defaults to 'name'.}

\item{status.col}{a vector with the name of the column with the
taxonomic status in the reference data frame 'df.ref'. Defaults to
'taxon.status'.}

\item{type.match.col}{a vector with the name of the column with the
type of name match. Defaults to 'match_type'.}

\item{mult.match.col}{a vector with the name of the column with the
information on if there was a unique or a multiple name match.
Defaults to "multiple_match"}

\item{mult.matches}{a character of how to deal with multiple
matches. Options are 'all' (concatenate all matches in the same
line) and 'best' (only one name match is returned per line).
Defaults to 'all'.}

\item{agg.cols}{a vector with the names of the columns that should
be concatenated in the case all matches are returned in the same
line (see argument 'mult.matches')}
}
\description{
Organize unique and multiple name matches in a reference
backbone for taxon names without author
}
\details{
To obtain the best match the column defined in the argument
'status.col' is used to organize the results and then remove all
duplicated names. The priority order to name with the best match
is (from high to low priority): 'accepted', 'synonym',
'unchecked', 'unplaced' and '' (i.e. empty taxonomic status).
}
\author{
Renato A. Ferreira de Lima
}
\keyword{internal}
