% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTDWG.R
\name{getTDWG}
\alias{getTDWG}
\title{Format People's Name To The TDWG Format}
\usage{
getTDWG(x, keep.prep = FALSE)
}
\arguments{
\item{x}{the character string with a name.}

\item{keep.prep}{logical. Should name prepositions be returned? Default to
FALSE.}
}
\value{
The character string \code{x} in the TDWG format.
}
\description{
Convert people's names from the long format (i.e. 'First Name(s)
  + Last Name(s)') to the \href{https://www.tdwg.org/}{Biodiversity Information Standards}
  (TDWG) format (i.e. Last Name, Initials of First Names(s)).
}
\details{
The standard name format suggested by the
  \href{https://www.tdwg.org/}{TDWG} is: Last name, followed by a comma and
  then the initials, separated by points (e.g. Hatschbach, G.G.).

  The function implicitly assumes that the last name is the one at the end of
  the name string.

  The function identifies common name prefixes or prepositions (e.g. de,
  dos, van, ter, ...). By default, these prefixes and prepositions are
  removed, but they can be returned if the argument `keep.prep` is set to
  TRUE.

  The functions uses internally another __plantR__ function: `lastName()`.
  So, the function deals with simples last names, as well as with compound
  last names and last names with generational suffixes.

  The function asumes that all names already containig commas are in the TDWG
  format. So, this cases are not edited. If only one name is given, the fucntion return
  the same name with the first letter capitalized. The function currently does not handle
  multiple names within the same name string. The function output it is
  relatively stable regarding the input format, lower/uppercasing and spacing, but it may not
  work in all cases (see examples below).
}
\examples{
  # Simple names
  getTDWG("gentry")
  getTDWG("Al Gentry")
  getTDWG("Alwyn Howard Gentry")
  getTDWG("Gert G. Hatschbach")
  getTDWG("G. G. HATSCHBACH")

  # Names already in the TDWG format (does nothing)
  getTDWG("Gentry, AH")
  getTDWG("Gentry, A.")

  # Name with prepositions
  getTDWG("Carl Friedrich Philipp von Martius")
  getTDWG("Alphonse Louis Pierre Pyrame de Candolle")
  getTDWG("Simon Jan van Ooststroom")
  getTDWG("Jan van der Hoeven")
  getTDWG("Maria da Silva")
  getTDWG("Maria da Silva", keep.prep = TRUE)

  # Names with generational suffixes
  getTDWG("Hermogenes Leitao Filho")
  getTDWG("J.E.Q. Faria Junior")
  getTDWG("S.J. Filho Neto")

  # Compound last name (needs to be marked with a '-')
  getTDWG("Augustin Saint-Hilaire")
  getTDWG("a. saint-hilaire")
  getTDWG("Saint-Hilaire A.")
  getTDWG("Augustin Saint Hilaire") # compound name missing '-'

  # Unusual formatting (function won't work always...)
  getTDWG("L. McDade") #cannot recognize names starting with Mc or Mac
  # two or more names: output incorrect (combine names of authors)
  getTDWG("C. Mendonca Filho; F. da Silva")


}
\references{
Conn, Barry J. (ed.) (1996). HISPID 3 - Herbarium Information Standards and
  Protocols for Interchange of Data. Herbarium Information Systems Committee'
  (HISCOM). https://www.tdwg.org/standards/hispid3/

Willemse, L.P., van Welzen, P.C. & Mols, J.B. (2008).
  Standardisation in data-entry across databases: Avoiding Babylonian
  confusion. Taxon 57(2): 343-345.
}
\author{
Renato A. F. de Lima
}
