% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYear.R
\name{getYear}
\alias{getYear}
\title{Get Year}
\usage{
getYear(x, noYear = "s.d.")
}
\arguments{
\item{x}{the character string}

\item{noYear}{Character. Standard for missing data in Year}
}
\value{
the character string \code{x}
}
\description{
Extracts the year of collection or determination from a field containing dates.
}
\details{
The function puts...
}
\examples{
# A vector with some typical examples of formats found in herbarium labels
dates = c("25/03/1981","25-03-1981","03/1981","03-1981","1981","1.981","1,1981",
"25/III/1981","25-III-1981","III/1981","III-1981","25031981","25 03 1981","'81",
"March 1981","Mar. 1981","n.d.","s.d.","s/d","",NA)
# Using the function to extract the year
getYear(dates)
# Using the function with a different character to indicate missing dates
getYear(dates, noYear = "n.d.")

#not working for
dates = c("31.12.1982","1982.31.12", '31/12')
getYear(dates, noYear = 'pouette')



}
\references{
Willemse, L.P., van Welzen, P.C. & Mols, J.B. (2008). Standardisation in data-entry across databases: Avoiding Babylonian confusion. Taxon 57(2): 343-345.
}
\author{
Lima, R.A.F.
}
