% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{islandsBuff}
\alias{islandsBuff}
\title{Buffer of World Minor Islands}
\format{
a MULTIPOLYGON 'sf' object with one class: land.
}
\source{
\url{https://www.naturalearthdata.com/downloads/10m-physical-vectors/10m-minor-islands/}
}
\usage{
data(islandsBuff)
}
\description{
Map used to support the validation of the original geographical coordinates.
It consists in a 0.25 degree buffer (~ 30 km in the Equator) around the minor
islands of the world (2 square km or less in size; scalerank >5). All buffers
were aggregated and the aggregated buffers were simplified (tolerance 0.01
decimal degrees). Note that buffers were generated at the largest scale
available in 'Natural Earth' (i.e. 1:10 m) and simplification was performed
after generating and aggregating the buffered polygons. Reference system:
WSG84
}
\keyword{datasets}
