% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landBuff}
\alias{landBuff}
\title{Buffer of World Land Map}
\format{
a MULTIPOLYGON 'sf' object with one class: land.
}
\source{
\url{https://www.naturalearthdata.com/downloads/10m-physical-vectors/10m-land/}
}
\usage{
data(landBuff)
}
\description{
Map used to support the validation of the original geographical coordinates.
It consists in a 0.5 degree buffer (~ 55 km in the Equator) around continents
and major islands (scalerank < 3) and a 0.25 degree buffer (~ 30 km in the
Equator) around medium islands (scalerank < 3). All buffers were then
aggregated and the aggregated buffers were simplified (tolerance 0.01 decimal
degrees). Note that buffers were generated at the largest scale available
in 'Natural Earth' (i.e. 1:10 m) and simplification was performed after
generating and aggregating the buffered polygons.
}
\details{
Reference system: WSG84
}
\keyword{datasets}
