% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_geo.R
\name{minDist}
\alias{minDist}
\title{Minimun Distance between Coordinates}
\usage{
minDist(lon, lat, min.dist = 0.001, output = NULL)
}
\arguments{
\item{lon}{numerical. Longitude in decimal degrees}

\item{lat}{numerical. Latitude in decimal degrees}

\item{min.dist}{numerical. Minimun threshold distance (in kilometers) to be
used to detect duplicated coordinates. Default to 1 meter.}

\item{output}{character. The type of information that should be returned (see Details)}
}
\value{
A vector of TRUE/FALSE or of minimun distances in kilometers.
}
\description{
The function calculates the minimum distance between coordinates
or the coordinates which are below a mininum threshold distance.
}
\details{
The argument \code{output} controls the type of output that should be
returned:
\itemize{
\item 'flag': a TRUE/FALSE vector indicating the coordinates which are below
\code{min.dist} from other coordinates.
\item 'group': the position of the first coordinate representing a group of
duplicated coordinates.
\item 'dist': the distance of each coordinate to the closest coordinated.
}
}
\examples{

lat = c(-23.475389, -23.475389, -23.475390, -23.475389)
lon = c(-47.123768, -47.123768, -47.123768, -47.123868)

minDist(lon, lat, output = 'group')
minDist(lon, lat, output = 'flag')
minDist(lon, lat, output = 'dist')

}
\seealso{
\link[plantR]{geoDist}
}
\author{
Renato A. F. de Lima
}
\keyword{internal}
