% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missName.R
\name{missName}
\alias{missName}
\title{Homogenize Notation For Missing Name}
\usage{
missName(x, type = NULL, noName = "Anonymous")
}
\arguments{
\item{x}{the character string.}

\item{type}{type of name: 'collector' or 'identificator'.}

\item{noName}{standard notation for missing names in \code{x}.}
}
\value{
the character string \code{x} in the fixed TDWG format.
}
\description{
Standardize the different notation of missing collector or
  identifier names
}
\examples{
# Needs to chose between 'collector' or 'identificator'
 missName(c("Gentry, AH", "s/col.", NA, "?", "s/c", "s/coletor"),
           type = "collector",
           noName = "s./c.")

 missName(c("Gentry, AH", "s/col.", NA, "?", "s/c", "s/coletor"),
         type = "collector",
         noName = "Anonymous")

 missName(c("Gentry, AH", "s/det.", "s/det", "s/d", "Determiner unknown"),
          type = "identificator",
          noName = "Anonymous")

}
\references{
Conn, Barry J. (ed.) (1996). HISPID 3 - Herbarium Information Standards and
  Protocols for Interchange of Data. Herbarium Information Systems Committee'
  (HISCOM). https://www.tdwg.org/standards/hispid3/
}
\author{
Renato A. F. de Lima
}
