% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoord.R
\name{plotCoord}
\alias{plotCoord}
\title{Plot Validation of Geographical Coordinates}
\usage{
plotCoord(
  x,
  orig.lon = "decimalLongitude",
  orig.lat = "decimalLatitude",
  new.lon = "decimalLongitude.new",
  new.lat = "decimalLatitude.new",
  geo.check = "geo.check"
)
}
\arguments{
\item{x}{a data frame with the occurrence data and the columns containing the
outputs of the \strong{plantR} validation functions.}

\item{orig.lon}{character. Name of the column with the original longitude.
Default to 'decimalLongitude.new'}

\item{orig.lat}{character. Name of the column with the original latitude.
Default to 'decimalLatitude.new'}

\item{new.lon}{character. Name of the column with the edited/validated
longitude. Default to 'decimalLongitude.new'}

\item{new.lat}{character. Name of the column with the edited/validated
latitude. Default to 'decimalLatitude.new'}

\item{geo.check}{character. Column with the result from the coordinate
checking. Default to 'geo.check'}
}
\description{
This function plots the comparison between the geographical
coordinates of records before and after their validation
}
\details{
This is a simple function to visually inspect the results of the
\strong{plantR} routines to standardize, retrieve missing coordinates and
validate the original ones.
}
\seealso{
\link[plantR]{prepCoord}, \link[plantR]{getCoord},
\link[plantR]{checkCoord}, , \link[plantR]{checkOut}, \link[plantR]{getCult}
}
\keyword{internal}
