% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepCoord.R
\name{prepCoord}
\alias{prepCoord}
\title{Format Geographical Coordinates}
\usage{
prepCoord(x)
}
\arguments{
\item{x}{a data.frame containing the geographical coordinates.}
}
\value{
x with the edited geographical coordinates.
}
\description{

}
\details{
For the time being the function does not differentiate missing coordinates
from those provided in a format that `prepCoord` does not handle. It also does not
differentiate original coordinates in decimal degrees in a good format from those
that needed to be re-formatted.
}
\examples{

}
\author{
Lima, R.A.F.
}
