% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepDup.R
\name{prepDup}
\alias{prepDup}
\title{Prepare For Duplicate Specimen Search}
\usage{
prepDup(
  x,
  noYear = "s.d.",
  noName = "s.n.",
  noNumb = "s.n.",
  comb.fields = list(c("family", "col.last.name", "col.number", "municipality"),
    c("family", "col.year", "col.number", "municipality"), c("species", "col.last.name",
    "col.number", "col.year"), c("col.year", "col.last.name", "col.number",
    "municipality")),
  ignore.miss = TRUE,
  ignore.na = TRUE,
  ...
)
}
\arguments{
\item{x}{a data frame with the species records.}

\item{noYear}{Character. Standard for missing data in Year. Default to
"n.d.".}

\item{noName}{character. Standard for missing data in collector name. Default
to "s.n.".}

\item{noNumb}{character. Standard for missing data in collector number. Default
to "s.n.".}

\item{comb.fields}{list. A list containing a vector with the occurrence
information that should be used to construct the duplicate search strings.
Default to TRUE. See Details.}

\item{ignore.miss}{logical. Should the duplicate search strings with
missing data (e.g. 'n.d.', 's.n.') be excluded from the duplicate search.
Default to TRUE.}

\item{ignore.na}{character. Should the duplicate search strings with
missing data (e.g. 'n.d.', 's.n.') be excluded from the duplicate search.
Default to TRUE.}
}
\description{
This function creates the duplicate search strings which are 
 created by combining information on the taxonomy, collection and locality 
 of the records.
}
\details{
The input data frame \code{x} must contain at least the columns ...

Three groups of fields are available to produce the duplicate search string,
and they are related to taxonomy, collection and locality of the specimen.
These fields are: 'family', 'species', 'col.name', 'col.last.name',
'col.number', 'col.year', 'municipality' or 'locality'. Each combination of
those fields (e.g. 'col.name' and 'col.number') should be provided as a vector and
as a separate element of this list. See examples.
}
\author{
Renato A. F. de Lima
}
