% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepLoc.R
\name{prepLoc}
\alias{prepLoc}
\title{Format Locality String}
\usage{
prepLoc(x)
}
\arguments{
\item{x}{the concatenated string of localities to be edited.}
}
\value{
the edited locality strings.
}
\description{
Edit and simplify the notation plantR locality search string
}
\details{
The function removes unwanted hifens and prepositions of locality names
}
\examples{
  prepLoc("Brazil_Rio de Janeiro")
  prepLoc("Brazil_São Paulo_São José do Rio Preto")
  prepLoc("Brazil_Bahia_Arraial d'Ajuda")
  prepLoc("Brazil_São Paulo_Sete Barras_Parque Estadual de Carlos Botelho")

}
\author{
Lima, R.A.F.
}
