% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepState.R
\name{prepState}
\alias{prepState}
\title{Format State Names}
\usage{
prepState(
  x,
  state.name = "stateProvince",
  country.name = "country",
  to.lower = TRUE,
  special.char = FALSE,
  rm.abbrev = TRUE
)
}
\arguments{
\item{x}{a vector of a data frame with state names to be standardized}

\item{state.name}{character. The name of the column containing the
state information. Defaults to "stateProvince" following the DWC
standard.}

\item{country.name}{character. The name of the column containing the
country information. Defaults to "country" following the DWC
standard.}

\item{to.lower}{logical. Should the output names be return in lower
cases? Default to TRUE.}

\item{special.char}{logical. Should special characters be
maintained? Default to FALSE.}

\item{rm.abbrev}{logical. Should common name abbreviation be
replaced? Default to TRUE.}
}
\value{
The input vector \code{x} in the standard name notation
(see Details)
}
\description{
Simple function to standardize the notation of state,
department and/or province name (administrative level 1) by
converting codes to their long names and by removing special
characters and some prepositions and separators.
}
\details{
State information is formatted into a standard notation
require by \strong{plantR}. Typical names of localitite at the
adm-level 1 (e.g. state, department, province) are also removed.
By default, all letters are lower-cased (argument \code{to.lower}) and
special characters (argument \code{special.char}) and common
abbreviations (e.g. 'st.') are replaced (argument \code{rm.abbrev}).
These edits aim at reducing possible variation in state name
notation and facilitate further data processing and comparison
within the \strong{plantR} workflow, increasing the chances of match
against the package internal gazetteer.

All state information with less than four letters are treated as
state codes or abbeviations and they are converted, if possible,
to the long name format.
}
\examples{
# Creating a data frame with locality information
estados <- c("RJ", "Rio de Janeiro", "MG", "estado de Minas Gerais",
             "Minas Gerais state", "state of Minas Gerais", "Minas Geraes",
             NA, "", "Minas Gerais", "MI")
paises <- c(rep("Brazil", 9), "", "Argentina")
df <- data.frame(country = paises, stateProvince = estados)

# Formating the locality information
prepState(df)
prepState(df, to.lower = FALSE)
prepState(df, rm.abbrev = FALSE)

}
\author{
Renato A. Ferreira de Lima
}
