% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepTDWG.R
\name{prepTDWG}
\alias{prepTDWG}
\title{Format People's Name In TDWG Format}
\usage{
prepTDWG(x, sep = ", ", format = "last_init", get.prep = FALSE)
}
\arguments{
\item{x}{the character string or vector.}

\item{sep}{character. Separator of the name format. Default to ", ".}

\item{format}{character. Output name format. The default ("last_init") is the
similar to the TDWG standard, but the inverse format can also be used
("init_last").}

\item{get.prep}{logical. Should last name prepositions be included? Default to
FALSE.}
}
\value{
The character string \code{x} in the fixed TDWG format.
}
\description{
Solve minor formatting issues of names that are already provided
  in the \href{https://www.tdwg.org/}{Biodiversity Information Standards}
  (TDWG) format.
}
\details{
The function fixes simple problems (e.g. missing points between name
  initials) for names provided already in a the following format: last name(s)
  + comma + name or initials.

  If the name provided is not separated by commas (the default separator),
  the function returns the same input character (i.e. no formatting is
  performed). In case of vector, only the names separated by commas will
  be edited.

  If the name provided has unusual formatting or if multiple names are
  provided within the same string, the function may not work properly. So,
  the output may depend on the input format and some level of
  double-checking may be necessary. See examples below.
}
\examples{
  # Simple name
  prepTDWG("Gentry, A.H.") # supposed not to change
  prepTDWG("Gentry, AH")
  prepTDWG("GENTRY, A H")
  prepTDWG("Gentry, Alwyn")
  prepTDWG("Gentry, Alwyn Howard")
  prepTDWG("gentry, alwyn howard")
  prepTDWG("Gentry, A.h.")
  prepTDWG("gentry, a.")
  prepTDWG("gentry, a.h.")
  prepTDWG("gentry, a. h.")
  prepTDWG("gentry, a")

  # Name not in TDWG format (no comma; does nothing)
  prepTDWG("Gentry AH")
  prepTDWG("Dardano de Andrade-Lima")

  # Names with generational suffixes
  prepTDWG("Leitao Filho, HF")
  prepTDWG("Filho Neto, S J")
  prepTDWG("Leitao filho, H. F.")

  # Compound last name
  prepTDWG("Saint-Hilaire, Augustin")
  prepTDWG("Ziffer Berger, Richter")
  prepTDWG("Saint-hilaire, a.")

  # Names with prepositions
  prepTDWG("Silva, Maria A. Pereira da")
  prepTDWG("Silva, Maria A. Pereira Da")
  prepTDWG("da Silva, Maria A. Pereira")
  prepTDWG("Silva, Maria A. Pereira da", get.prep = TRUE)
  prepTDWG("Silva, Maria A. Pereira Da", get.prep = TRUE)
  prepTDWG("da Silva, Maria A. Pereira", get.prep = TRUE)

  # Multiple names
  names <- c("Gentry, AH", "Gentry A.H.",
  "Leitao filho, H. F.", "Saint-Hilaire, a",
  "gentry, alwyn howard", "da Silva, Maria A.")
  prepTDWG(names)
  prepTDWG(names, format = "init_last")
  prepTDWG(names, format = "init_last", get.prep = TRUE)

  # Unusual formatting (function won't work always...)
  # one name, two commas: fails to get all names
  prepTDWG("Cesar Sandro, Esteves, F")
  # two or more names: output incorrect (combine names of authors)
  prepTDWG("Mendonca Filho, C.V.; Neto, F.C.C.")
  # two or more names, separated by comma: output incorrect (combine names of authors)
  prepTDWG("A. Alvarez, A. Zamora & V. Huaraca")

}
\references{
Conn, Barry J. (ed.) (1996). HISPID 3 - Herbarium Information Standards and
  Protocols for Interchange of Data. Herbarium Information Systems Committee'
  (HISCOM). https://www.tdwg.org/standards/hispid3/

Willemse, L.P., van Welzen, P.C. & Mols, J.B. (2008).
  Standardisation in data-entry across databases: Avoiding Babylonian
  confusion. Taxon 57(2): 343-345.
}
\author{
Renato A. F. de Lima
}
