% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptr_search.R
\name{ptr_search}
\alias{ptr_search}
\title{Search the USDA PLANTS database API}
\usage{
ptr_search(
  symbol = NULL,
  scientific_name = NULL,
  common_name = NULL,
  state_province = NULL,
  county = NULL,
  nativityIntroducedSelected = NULL,
  nativityNativeSelected = NULL,
  federal_noxious = NULL,
  state_noxious = NULL,
  federal_te = NULL,
  state_te = NULL,
  basic = TRUE,
  flatten = TRUE
)
}
\arguments{
\item{symbol}{Returns plants that match the USDA symbol given. Example
values: ABBA, CAR.}

\item{scientific_name}{Returns plants that match the specified scientific
name (without author).  Parameter is case sensitive.
Example value: Abies balsamea.}

\item{common_name}{Represents the National Common Name (Also known as the
accepted or preferred common name). Returns plants that match the specified
common name. Parameter is case sensitive. Example values: balsam fir.}

\item{state_province}{Returns plants that are present in the specified state
or province. When state is specified it should be referenced as state name.
Example values: Alabama, Colorado, Alberta.}

\item{county}{Returns plants that are present in the specified county.
Example values (for the state of Colorado): Adams, Larimer, Yuma.}

\item{nativityIntroducedSelected, nativityNativeSelected}{Only one of these
parameters may be specified. A list of acceptable values is included below.}

\item{federal_noxious}{Returns only plants that have the specified federal
noxious status. Current statuses are "Noxious Weed" (\code{nw}), "Quarantine"
(\code{q}).}

\item{state_noxious}{Returns only plants that are designated as noxious weeds
(or equivalent) in a specified state. As the categories of noxious weeds vary
greatly by state, specific noxious codes are not used as a filter. When state
is specified it can be referenced as state code. Example values: UT, AZ}

\item{federal_te}{Returns only plants that have the specified federal
threatened, endangered status. The possible statuses are \code{endangered} or
\code{threatened}.}

\item{state_te}{Returns only plants that are designated on any threatened or
endangered list in the specified state. As the categories of threatened and
endangered weeds vary greatly by state, specific codes are not used as a
filter. When state is specified it can be referenced as state code. Example
values: CO, CA}

\item{basic}{Logical: should the search use the basic API (returning data on
plant symbols, accepted common names, and scientific names) or the full
API which returns much more data? Note that queries using the full API
without many parameters (such as searching using a common name or symbol
without specifying restrictions using the other arguments) may take a long
time, time out or error in unexpected ways, or (if more than 10 species will
be returned) fall back to using the basic API with a warning. When using the
full API, it is often a good idea to run multiple queries for smaller regions
rather than attempting to run a single query for a larger region.}

\item{flatten}{Logical: attempt to flatten nested data frames into a
data frame with fewer nested layers? Passed to \link[jsonlite:fromJSON]{jsonlite::fromJSON}.}
}
\value{
A data frame containing the data returned by the API call requested.
}
\description{
Most parameter descriptions below have been adapted from the official USDA
PLANTS Web Services Interface Definition.
If the search is broad, and if the number of symbols to be returned are
greater than 10,then only basic information of those symbols are returned.
If more detailed records are required, please narrow the search criteria.
}
\section{Nativity status}{

Only one of \code{nativityIntroducedSelected} and \code{nativityNativeSelected} may be
specified in a single API call, as filtering for only introduced species in a
region necessarily excludes native species in that region.

Valid values for these fields are:
\itemize{
\item \code{l48}: Lower 48 states
\item \code{Ak}: Alaska
\item \code{Hi}: Hawaii
\item \code{pr}: Puerto Rico
\item \code{vi}: U.S. Virgin Islands
\item \code{can}: Canada
\item \code{gl}: Greenland (Denmark)
\item \code{spm}: St. Pierre and Miquelon (France)
\item \code{na}: North America
}
}

\examples{
\dontrun{
# Get basic results matching the common name "balsam fir":
ptr_search(common_name = "balsam fir")

# Get full results for the same query:
ptr_search(common_name = "balsam fir", basic = FALSE)

}

}
