% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjabot.R
\name{rjabot}
\alias{rjabot}
\title{Gets occurrence data from jabot}
\usage{
rjabot(dir = "results/", filename = "output", species = NULL)
}
\arguments{
\item{dir}{Path to directory where the file will be saved. Default is to create a "results/" directory}

\item{filename}{Name of the output file}

\item{species}{Genus and epithet separated by space. More than one should be concatenated in a vector}
}
\value{
A list of two elements. The first element is a character string containing the url search and the second element is a data.frame with the search result. It also saves the output on disk
}
\description{
This function gets occurrence data in Jabot's API (http://servicos.jbrj.gov.br/jabot/). Providing a scientific name for a species is mandatory.
}
\author{
Sara Mortara
}
