% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmLatin.R
\name{rmLatin}
\alias{rmLatin}
\title{Remove Special Characters}
\usage{
rmLatin(x)
}
\arguments{
\item{x}{a vector with characters to be replaced}
}
\value{
the input vector with the special characters replaced
}
\description{
Simple function to replace accents by the same or corresponding
letters without accents. Currently, the function basically replaces special
characters corresponding to the Latin-1 Supplement and very few Latin
Extended-A. For a more complete replacement of all non-ASCII characters,
please use the function \code{replace_non_ascii()} from package \strong{textclean}.
}
\examples{
  nomes <- c("Thomé", "Muñoz", "œuf")
  Encoding(nomes) <- "latin1"
  nomes

  rmLatin(nomes)

}
\author{
Renato A. F. de Lima
}
\keyword{internal}
