% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspeciesLink.R
\name{rspeciesLink}
\alias{rspeciesLink}
\title{Gets occurrence data from speciesLink}
\usage{
rspeciesLink(
  dir = "results/",
  filename = "output",
  save = FALSE,
  basisOfRecord = NULL,
  family = NULL,
  species = NULL,
  collectionCode = NULL,
  country = NULL,
  stateProvince = NULL,
  county = NULL,
  Coordinates = NULL,
  CoordinatesQuality = NULL,
  Scope = NULL,
  Synonyms = "no synomyms",
  Typus = FALSE,
  Images = NULL,
  RedList = FALSE,
  MaxRecords = NULL,
  file.format = "csv",
  compress = FALSE
)
}
\arguments{
\item{dir}{Path to directory where the file will be saved. Default is to
create a "results/" directory}

\item{filename}{Name of the output file}

\item{save}{Logical. Save output to filename? Defaults to FALSE}

\item{basisOfRecord}{Character. Any in 'PreservedSpecimen', 'LivingSpecimen',
'FossilSpecimen', 'HumanObservation', 'MachineObservation' or
'MaterialSample'. Default is 'PreservedSpecimen' for museum and herbarium
search}

\item{family}{Family name. More than one name should be concatenated in a
vector}

\item{species}{Genus or genus and epithet separated by space. More than one
should be concatenated in a vector}

\item{collectionCode}{Any collection available at speciesLink. Example: ALCB,
E, INPA, MOBOT_BR.  Accepts a vector of names}

\item{country}{Any country name. No ASCII characters allowed. Accepts a
vector of names}

\item{stateProvince}{Any state or province. No ASCII characters allowed.
Accepts a vector of names}

\item{county}{Any municipality name. No ASCII characters allowed. Accepts a
vector of names}

\item{Coordinates}{Specify if records should have coordinates. Default is
"no check" but it also accepts "Yes", No", "Original", "Automatic", "Blocked"
or "no check"}

\item{CoordinatesQuality}{Any character in "Good" or "Bad" to select specific
type of coordinates}

\item{Scope}{Group to be required. If NULL searches all groups. Any in
"plants", "animals", "microrganisms" or "fossils"}

\item{Synonyms}{If species names should be checked for synonyms in a specific
dictionary. Set to "species2000" for search in Catálogo da Vida species2000,
"flora2020" for Flora do Brasil 2020, "MycoBank" for MycoBank, "AlgaeBase"
for AlgaeBase, "DSMZ" for  DSMZ Prokaryotic Nomenclature Up-to-Date,
"Moure" for Catálogo de Abelhas Moure or "no synonyms".}

\item{Typus}{Logic. If TRUE select only typus}

\item{Images}{If select only records with images. Default is NULL.
It accepts: "Yes", "Live", "Polen", "Wood"}

\item{RedList}{Logic. If TRUE only species in the IUCN Red List are returned}

\item{MaxRecords}{Numeric. Maximum number of records to be required}

\item{file.format}{Character. The file extension to be used for saving ('csv'
or 'rds'). Default to 'csv'.}

\item{compress}{Logical. Should the file be compressed? Default to FALSE.}
}
\value{
A data.frame with the search result, which can also be saved on disk
}
\description{
This function access version beta 0.1 of speciesLink API
(https://api.splink.org.br/) and returns occurrence data of species.
}
\details{
The time necessary to download records from the speciesLink API will
depend on the number of records and species in the query and from the speed
of the internet connection.

The speciesLink API does not allow the download of 10 or more taxa at a time.
So to download records from larger lists of species, you will probably need
to make the queries in a loop.
}
\examples{
\dontrun{
ex01 <-
rspeciesLink(filename = "ex01",
                    species =  c("Eugenia platyphylla", "Chaetocalyx acutifolia"),
                    Scope = "plants")
}

}
\author{
Sara Mortara
}
