% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBorders.R
\name{shares_border}
\alias{shares_border}
\title{Checks if two countries share a border}
\usage{
shares_border(country1 = "brazil", country2 = "argentina")
}
\arguments{
\item{country1}{First country to check, use spData standard}

\item{country2}{Second country to check, use spData standard}
}
\description{
Checks if two countries share a border
}
\author{
Andrea Sánchez-Tapia & Sara Mortara
}
\keyword{internal}
