% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shoreLines}
\alias{shoreLines}
\title{World Shore Lines}
\format{
a MULTILINESTRING 'sf' object.
}
\source{
\url{https://www.naturalearthdata.com/downloads/50m-physical-vectors/50m-land/}
}
\usage{
data(shoreLines)
}
\description{
Map used to support the validation of the original geographical coordinates.
It was generated from the map available in 'Natural Earth' at medium scale
(i.e. 1:50 m), which includes continents and major islands. Original polygons
of each country were aggregated, then simplified (tolerance 0.001 decimal
degrees) and finally converted into 'spatial lines'. For some reason this
procedure generated objects smaller in size than the 'coastline' maps
available in 'Natural Earth'. Reference system: WSG84
}
\keyword{datasets}
