% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryData.R
\name{summaryData}
\alias{summaryData}
\title{Summary of Species Occurrence Data}
\usage{
summaryData(x, print = TRUE, top = 5)
}
\arguments{
\item{x}{a data frame with the occurrence data.}

\item{print}{logical. Should the first part of the table be printed? Default
to TRUE.}

\item{top}{numerical. Number of groups (e.g. collections, families) to be
printed. Default to 5.}
}
\description{
This function provides a general summary of the occurrence data,
including the number of records, biological collections, taxa and
countries.
}
\details{
The summary output depends on the presence of some key columns in
the input data frame \code{x}, which should follow the Darwin Core
standards. If the edited columns from \strong{plantR} are given, they are given
priority to report the data summaries.

The function prints the summary tables related to the occurrence data.
However, the tables generating those summaries can be saved into an object.
}
