% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdwgNames.R
\name{tdwgNames}
\alias{tdwgNames}
\title{Format Multiple People's Names To TDWG Standard}
\usage{
tdwgNames(
  x,
  sep.in = c(";", "&", " e ", " et ", "\\\\|"),
  sep.out = "|",
  out = c("all")
)
}
\arguments{
\item{x}{the character string.}

\item{sep.in}{a vector with the symbols separating multiple names in the input
string. Default to any of the following: ";", "&", " e ", " et " and "|"}

\item{sep.out}{a character string with the symbol separating multiple names in
the output string. Defaults to " | ".}

\item{out}{a character string with the type of output desired: all names,
first name, or auxiliary names.}
}
\value{
The character string \code{x} in the TDWG format (Last Name {comma}
First Name). If the names of only one person is given, the function returns
the same as function \code{tdwgName}.
}
\description{
Convert multiple collector or determiner names into the
\href{https://www.tdwg.org/}{Biodiversity Information Standards} (TDWG).
}
\details{
The function puts the name of multiple persons into the format as
suggested by the \href{https://www.tdwg.org/}{Biodiversity Information
Standards} (TDWG). The standard name format is: last name, followed by a comma
and then the initials, separated by points (e.g. Hatschbach, G.G.).

The function is relatively stable regarding the input format and spacing,
but it may not work in all cases, particularly if the string provided
already contain commas. See examples below.

Currently, the function removes name prefixes or prepositions (e.g. de,
dos, van, ter, ...). Also, it removes some titles (i.e. Dr., Prof., Pe.)
but not all of them (e.g. Doctor, Priest, etc.). The function also does not
handle hyphenated first names.

The user must choose the input character separating the multiple names. By
default, the function uses ';', ' e ', ' et ' '&' and '|' as. Users can also
choose the output separator between names, but the default is the TDWG
recommendation, which is " | ".

If special separators are used (e.g. '.,'), make sure that their notation
contains the scape symbols required when using regular expressions in R
(e.g. '\\\.,').

The user can also choose the output character to be all names in the input
character (out = "all"), only the name (out = "first") or only the
auxiliary/secondary names (out = "aux"). By default, the function returns
all names together, as recommended by the TDWG standards.
}
\examples{
  # Simple names
  tdwgNames("Karl Emrich & Balduino Rambo")
  tdwgNames("R. Reitz; R.M. Klein")
  tdwgNames("Raulino Reitz; R.M. Klein", sep.out= " & ")
  tdwgNames("Alwyn H. Gentry|Elsa M. Zardini")
  tdwgNames("Gert G. Hatschbach")

  # Names with generational suffixes
  tdwgNames("Leitão Filho, H.F.; Shepherd, G.J.")
  tdwgNames("Leitao Filho, HF") ## CHECK!

  # Compound last names
  tdwgNames("A. Ducke; Dárdano de Andrade-Lima")

  # Multiple names separated by commas
  tdwgNames("A. Alvarez; A. Zamora & V. Huaraca")
  tdwgNames("A. Alvarez; A. Zamora & V. Huaraca", out = "first")
  tdwgNames("A. Alvarez; A. Zamora & V. Huaraca", out = "aux")
  tdwgNames("A. Alvarez; A. Zamora & V. Huaraca", out = "aux", sep.out = ";")

  # Multiple names separated by commas (does not always work)
  tdwgNames("A. Alvarez, A. Zamora & V. Huaraca") # output incorrect (combine names of authors)
  tdwgNames("A. Alvarez, A. Zamora & V. Huaraca", sep.in=c(",","&")) # output correct
  # output incorrect (make initials separate names)
  tdwgNames("Alvarez, A., Zamora, A. & Huaraca, V.", sep.in=c(",","&"))
  tdwgNames("Alvarez, A., Zamora, A. & Huaraca, V.", sep.in=c("\\\\.,","&")) # output correct
  # Unusual formatting (function won't always work)
  tdwgNames("Cesar Sandro, Esteves, F") # one name, two commas: fails to get the right last name

}
\references{
Conn, Barry J. (ed.) (1996). HISPID 3 - Herbarium Information Standards and
Protocols for Interchange of Data. Herbarium Information Systems Committee'
(HISCOM). https://www.tdwg.org/standards/hispid3/

Willemse, L.P., van Welzen, P.C. & Mols, J.B. (2008).
Standardisation in data-entry across databases: Avoiding Babylonian
confusion. Taxon 57(2): 343-345.
}
\seealso{
\link[plantR]{tdwgName}
}
\author{
Renato A. F. de Lima
}
