% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPL1.R
\name{TPL1}
\alias{TPL1}
\title{Standardize plant names according to The Plant List}
\usage{
TPL1(
  splist,
  genus = NULL,
  species = NULL,
  infrasp = NULL,
  infra = TRUE,
  corr = TRUE,
  diffchar = 2,
  max.distance = 1,
  version = "1.1",
  encoding = "UTF-8",
  author = TRUE,
  drop.lower.level = FALSE,
  file = "",
  silent = TRUE,
  repeats = 6
)
}
\arguments{
\item{splist}{A character vector specifying the input taxa, each element
including genus and specific epithet and, potentially, author name and
infraspecific abbreviation and epithet.}

\item{genus}{A character vector containing the genera of plant taxon names.}

\item{species}{A character vector containing the specific epithets of plant
taxon names.}

\item{infrasp}{A character vector containing the infraspecific epithets of
plant taxon names.}

\item{infra}{Logical. If TRUE (default), infraspecific epithets are used to
match taxon names in TPL.}

\item{corr}{Logical. If TRUE (default), spelling errors are corrected (only)
in the specific and infraspecific epithets prior to taxonomic
standardization.}

\item{diffchar}{A number indicating the maximum difference between the number
of characters in corrected and original taxon names. Not used if corr =
FALSE.}

\item{max.distance}{A number indicating the maximum distance allowed for a
match in agrep when performing corrections of spelling errors in specific
epithets. Not used if corr = FALSE.}

\item{version}{A character vector indicating whether to connect to the newest
version of TPL (1.1) or to the older one (1.0). Defaults to "1.1".}

\item{encoding}{Encoding to be assumed for input strings from TPL website;
defaults to "UTF-8" (see read.csv and Encoding).}

\item{author}{Logical. If TRUE (default), the function tries to extract
author names from the input taxon (see Details).}

\item{drop.lower.level}{Logical. If TRUE, the variety is dropped from the
input taxon if both subspecies and variety are given, and the forma is
dropped from the input taxon if both subspecies or variety and forma are
given. If specific and subspecific epithet are identical, the subspecies
(variety) part is dropped instead, and variety (forma) is kept. Defaults to
FALSE.}

\item{file}{Either a character string naming a file or a connection open for
writing. "" (default) indicates output to the console.}

\item{silent}{Logical. If FALSE, the function prints the taxon name that is
currently processed in the output. Defaults to TRUE.}

\item{repeats}{number indicating how many times TPLck should be called if no
connection to TPL website can be established (temporarily).}

\item{n}{a index.}
}
\description{
Edited copy of the function \code{TPL()} from package Taxonstand to
avoid an error.
}
\keyword{internal}
