% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateCoord.R
\name{validateCoord}
\alias{validateCoord}
\title{Workflow for geographical coordinates validation}
\usage{
validateCoord(
  x,
  lon = "decimalLongitude.new",
  lat = "decimalLatitude.new",
  country_shape = "NAME_0",
  country_gazetteer = "country"
)
}
\arguments{
\item{x}{Data.frame with coordinates in decimal degrees.}

\item{lon}{Column with the longitude to be validated}

\item{lat}{Column with the latitude to be validated}

\item{country_shape}{Name of the column with the country that comes from the shapefile}

\item{country_gazetteer}{Name of the column with the country that comes from the gazetteer}
}
\description{
This function performs the workflow for geographical coordinates validation
}
\author{
Andrea Sánchez-Tapia & Sara R. Mortara
}
