% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateDup.R
\name{validateDup}
\alias{validateDup}
\title{Prepare, Search and Merge Duplicate Specimens}
\usage{
validateDup(occ.df, merge = TRUE, remove = FALSE, ...)
}
\arguments{
\item{occ.df}{a data frame, containing typical fields from occurrence records from
herbarium specimens}

\item{merge}{logical. Should duplicates be merged? Default to TRUE.}

\item{remove}{logical. Should duplicates be removed? Default to FALSE.}

\item{...}{Parameters from mergeDup}
}
\value{
The input data frame \code{x}, plus the new columns with the formatted
  fields.
}
\description{
This function search for duplicated specimens within and across
  collections and it can be used to homogenize the information of
  different groups of fields and to remove duplicates, leaving only one
  occurrence for each group of duplicata.
}
\details{
The function works similarly to a wrapper function, where the
  individuals steps of the proposed __plantR__ workflow for preparing,
  searching, merging and removal of duplicates are performed altogether (see
  the __plantR__ tutorial for details).
}
\seealso{
\link[plantR]{prepDup}, \link[plantR]{getDup}, \link[plantR]{mergeDup},
 and \link[plantR]{rmDup}.
}
\author{
Renato A. F. de Lima
}
