% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workCoord.R
\name{workCoord}
\alias{workCoord}
\title{Get Working Geographical Coordinates}
\usage{
workCoord(x)
}
\arguments{
\item{x}{a data.frame containing the original coordinates and the coordinates obtained from a
gazetteer.}
}
\value{
x with the columns 'decimalLatitude.new' and 'decimalLongitude.new' filled with the
geographical coordinates of the gazetteer at the best resolution (i.e. administrative level) available.
}
\description{

}
\details{

}
\examples{

}
\author{
Lima, R.A.F.
}
