% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world}
\alias{world}
\title{Simplified World Map}
\format{
a MULTIPOLYGON 'sf' object with 255 features and 2 fields: iso_a2
(country code) and 'name' (country name).
}
\source{
\url{https://www.naturalearthdata.com/downloads/10m-cultural-vectors/10m-admin-0-countries/}
}
\usage{
data(world)
}
\description{
Map used to support the validation of the original geographical coordinates,
mainly regarding the detection of problematic coordinates fallign just
outside its country corders. It was generated from the best scale available
in 'Natural Earth' (i.e. 1:10 m), which consists in 255 polygons for 247
countries in the world. Country long names were edited to match the
\strong{plantR} default gazetteer and original polygons were simplified (tolerance
0.001 decimal degrees). Only the ISO 3166-1 alpha-2 codes and the long names
of each country are stored with the polygons. Reference system: WSG84
}
\keyword{datasets}
