% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildName.R
\name{buildName}
\alias{buildName}
\title{Build Taxon Name}
\usage{
buildName(x, col.names = c("genus", "specificEpithet"))
}
\arguments{
\item{x}{a data frame with the name information}

\item{col.names}{the name of the columns containing the information
to be combined in the desired order of output}
}
\value{
a vector with the combined name information in \code{col.names}
}
\description{
Combine different columns with species name
information (i.e. genus, epiteth, infra-epiteth, author) into a
single taxon name
}
\examples{
\dontrun{
df <- data.frame(genus = c("Aa", "Ab"), specificEpithet = c("aa", "bb"))
buildName(df)
}

}
\author{
Renato A. F. de Lima
}
\keyword{internal}
