% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capName.R
\name{capName}
\alias{capName}
\title{Capitalize Name}
\usage{
capName(x)
}
\arguments{
\item{x}{the character string to be capitalized.}
}
\value{
the character string equal to \code{x} with the first letter capitalized.
}
\description{
Capitalize the name of the collector or determiner of a biological specimen.
}
\details{
The function works for simple names and compound names which are separated by
a space or by the character '-'. It also works for names that are entirely capitalized.
}
\examples{
# Simple name
  capName("gentry")
  capName("HATSCHBACH")
# Names with generational suffixes
  capName("leitao filho")
# Compound name
  capName("saint-hilaire")
}
