% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capName.R
\name{capName}
\alias{capName}
\title{Capitalize Name}
\usage{
capName(x)
}
\arguments{
\item{x}{a character string to be capitalized.}
}
\value{
a character string equal to \code{x} with the first letter capitalized.
}
\description{
Capitalize the name of the person that collected or identified a biological specimen.
}
\details{
The function works for simple names and coumpound names which are separated by
a space or by thecharacter '-'.
}
\examples{
# Simple name
  capName("gentry")
# Names with generational suffixes
  capName("leitao filho")
# Compound name
  capName("saint-hilaire")
}
