% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capName.R
\name{capName}
\alias{capName}
\title{Capitalize Name}
\usage{
capName(x)
}
\arguments{
\item{x}{the character string or vector to be capitalized.}
}
\value{
the character string equal to \code{x} with the first letter
  capitalized.
}
\description{
Capitalize the first letter of people's (last) names.
}
\details{
The function works for simple names and compound names that are
  separated by a space or by the character '-'. It also works for names that
  are entirely capitalized.

  This function was conceived to edit people's last names. So, the function
  only works for up to two names and it works only partially for names
  containing initials. For strings with more than two names, the function
  `str_to_title()` from package `stringr::str_to_title` can be used.
}
\examples{
# Simple names
  capName("gentry")
  capName("HATSCHBACH")
  capName(c("gentry", "HATSCHBACH"))

# Names with generational suffixes
  capName("leitao filho")

# Compound last names
  capName("saint-hilaire")

# Full names (does not work)
  capName("hermogenes leitao filho")
  capName("auguste saint-hilaire")

# Names with initials (does not work completely)
  capName("a.h. gentry")

}
\author{
Renato A. F. de Lima
}
