% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBorders.R
\name{checkBorders}
\alias{checkBorders}
\title{Flags border points}
\usage{
checkBorders(
  x = data.frame(occs),
  country_shape = "NAME_0",
  country_gazetteer = "country"
)
}
\arguments{
\item{x}{occurrence data frame}

\item{country_shape}{Name of the column with the country that comes from the shapefile}

\item{country_gazetteer}{Name of the column with the country that comes from the gazetteer}
}
\description{
Flags border points
}
\author{
Andrea Sánchez-Tapia & Sara Mortara
}
