% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklist.R
\name{checklist}
\alias{checklist}
\title{Create Species Checklist}
\usage{
checklist(
  x,
  fam.order = TRUE,
  n.vouch = 30,
  type = "short",
  rm.dup = TRUE,
  rank.type = 5,
  date.format = "\%d \%b \%Y"
)
}
\arguments{
\item{x}{a data frame with the occurrence data, generally as the output of the
__plantR__ validation functions.}

\item{fam.order}{logical. Should taxa be organized in alphabetical order
of families? Default to TRUE.}

\item{n.vouch}{numerical. Maximum number of vouchers to be listed per taxa.
Default to 30.}

\item{type}{character. The type of voucher list desired. Options are: 'short',
'selected' and 'list' (see details below).}

\item{rm.dup}{logical. Should duplicated specimens be removed prior to the
calculation of species summaries? Default to TRUE.}

\item{rank.type}{numerical. Value of the ranking for type specimens in order
to organizeorder and filter the voucher list. Default to 5.}

\item{date.format}{The desired format for the dates. Default to "\%d \%b \%Y".}
}
\description{
This function creates a checklist of the species contained
in the occurrence data, including a list of voucher specimens.
}
\details{
The checklist can be organized in alphabetic order by taxa or in
  alphabetic order by family and then by taxa within families (the
  default).

  By default, the checklist provides the number of records found and the
  overall taxonomic and geographic confidence level of the records (columns
  'tax.CL' and 'geo.CL'), if available. The taxonomic confidence level is the
  percentage of records with the identification flagged as 'high', while the
  geographic confidence level is the percentage of records with coordinates
  flagged as being validated at municipality or locality levels.

  The function also provide a list of vouchers, giving priority to type
  specimens and records with higher level of confidence in their
  identification. By default, the function provides up to 30 vouchers
  per taxa, but this number can be controlled using the argument `n.vouch`.

  The voucher list can be provided in the following output formats (the
  option 'list' is not implemented yet):
  \itemize{
  \item 'short': Collector name, Collector number (collections of deposit)
  \item 'selected': COUNTRY, stateProvince: municipality, Date, Collector name,
  Collector number (collections of deposit)
  \item 'list': Collector name, Collector number(s) (species code)
  }

  Note: although we provide a `date.format` argument, checks of other date
  formats other then the default are pending and so they may not work
  properly.
}
\examples{
(df <- data.frame(collectionCode = c("CRI","CRI","CRI","CRI"),
catalogNumber = c("3565","713","3073","15331"),
recordedBy = c("Rebelo, M.C.","Citadini-Zanette, V.",
"Santos, R.","Zapelini, I."),
recordNumber = c("s.n.","1063","11","s.n."),
year = c("1994","1990","1994","2020"),
family = c("Salicaceae","Salicaceae","Cannabaceae","Cannabaceae"),
scientificName = c("Casearia sylvestris","Casearia sylvestris",
"Trema micrantha","Trema micrantha"),
country = c("brazil","brazil","brazil","brazil"),
stateProvince = c("santa catarina","santa catarina",
"santa catarina","santa catarina"),
municipality = c("jaguaruna","orleans","icara",NA)))

checklist(df, rm.dup = FALSE)
checklist(df, rm.dup = FALSE, type = "selected")


}
