% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colNumber.R
\name{colNumber}
\alias{colNumber}
\title{Format Collector Number}
\usage{
colNumber(x, colCodes = NULL, noNumb = "s.n.")
}
\arguments{
\item{x}{the character string}

\item{colCodes}{a character string with the collection codes to be removed
from the collector number. Default to NULL.}

\item{noNumb}{character. The desired notation to identify abscence data in
Collector Number. Defaults to "s.n."}
}
\value{
a character vector with the same length of \code{x} with the edited
  collector numbers
}
\description{
The function edits the field 'collector number' of the
  herbarium labels
}
\details{
The function attempts to clean and standardize the information
contained in the field 'Collector Number', which is typical of herbarium
samples. It performs several edits such as removal of unnecessary spaces,
letters, and the replacement of missing information into a standardized notation
}
\examples{
# A vector with some typical examples of formats found in herbarium labels
numbers <- c("3467","3467 "," 3467","ALCB3467","Gentry 3467","ALCB-3467","3467a","3467A","3467 A",
"3467-A","PL671","57-685","685 - 4724","1-80","-4724","(3467)","(3467","3467)","32-3-77",
"s/n.","s.n.","s.n","",NA)

# Using the function defaults
colNumber(numbers)

# Using the function to remove the collection code from the collector number
colNumber(numbers, colCodes = c("ALCB","ESA"))

# Defining user-specific abbreviations for the specimens without collector number
colNumber(numbers, colCodes = c("ALCB","ESA"), noNumb = "n.a.")

}
\references{
Willemse, L.P., van Welzen, P.C. & Mols, J.B. (2008).
  Standardisation in data-entry across databases: Avoiding Babylonian
  confusion. Taxon 57(2): 343-345.
}
\author{
Renato A. F. de Lima
}
