% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colNumber.R
\name{colNumber}
\alias{colNumber}
\title{Format Collector Number}
\usage{
colNumber(x, colCodes = NULL, noNumb = "s.n.")
}
\arguments{
\item{x}{a character string or vector}

\item{colCodes}{a character string with the collection codes to be removed
from the collector number. Default to NULL.}

\item{noNumb}{character. The desired notation in the case of missing
collector number. Defaults to "s.n."}
}
\value{
a character vector with the same length of \code{x} with the edited
collector numbers.
}
\description{
The function standardizes the 'collector number' which is
typically associated with biological records.
}
\details{
The function performs several edits such as removal of unnecessary
spaces, letters, parentheses, and the replacement of missing information of
collector numbers into a standardized notation, defined by the argument
\code{noNumb}. Names of authors are automatically removed but not collection
codes. Zeros and strings without numbers are treated as missing
information.
}
\examples{
# A vector with some typical examples of formats found in herbarium labels
numbers <- c("3467", "3467 ", " 3467", "ALCB3467", "Gentry 3467",
"A. Gentry 3467", "Gentry, A. 3467", "ALCB-3467", "ALCB 3467", "3467a",
"3467A", "3467 A", "3467-A", "PL671", "57-685", "685 - 4724", "1-80",
"-4724", "(3467)", "(3467", "3467)", "32-3-77", "s/n.", "s.n.", "", NA)

# Using the function defaults
colNumber(numbers)

# Using the function to remove the collection code from the collector number
colNumber(numbers, colCodes = c("ALCB", "ESA"))

# Defining user-specific abbreviations for specimens without collector number
colNumber(numbers, colCodes = c("ALCB", "ESA"), noNumb = "n.a.")

}
\references{
Willemse, L.P., van Welzen, P.C. & Mols, J.B. (2008).
Standardisation in data-entry across databases: Avoiding Babylonian
confusion. Taxon 57(2): 343-345.
}
\author{
Renato A. F. de Lima
}
