% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example}
\alias{example}
\title{Dataset to illustrate the data editing and validation}
\format{
A data frame with around 18,000 rows and 26 variables
}
\source{
\url{https://specieslink.net/}
}
\usage{
data(example)
}
\description{
A dataset containing the retrieved records of the species
\emph{Trema micrantha} and \emph{Casearia sylvestris} downloaded from the
speciesLink network in Feb 2026. Data was filtered to contain only
preserved specimens and the columns most relevant to \code{plantR}
}
\keyword{datasets}
