% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example}
\alias{example}
\title{Dataset to illustrate the data editing and validation}
\format{
A data frame with 16,491 rows and 51 variables
}
\source{
\url{http://splink.cria.org.br}
}
\usage{
data(example)
}
\description{
A dataset containing the retrieved records of the species *Trema micrantha*
and *Casearia sylvestris* downloaded from the speciesLink network in June
2020.
}
\keyword{datasets}
