% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixCase.R
\name{fixCase}
\alias{fixCase}
\title{Standardize Species Name Casing}
\usage{
fixCase(x)
}
\arguments{
\item{x}{the character string or vector to be standardized.}
}
\value{
the character string equal to \code{x} with the first letter of the
genus capitalized and the first letter of the (infra-)specific epithet in
lowercase.
}
\description{
Capitalize the genus first letter and lowercase the
(infra-)specific epithet of species names.
}
\details{
The function try to fix some common upper/lowercase issues in
species names, taking into account the possible presence of name authorities,
as an alternative to the function \code{fixCase()} from the package \code{flora}.

It works for names at any taxonomic level (e.g. genus, species and
infra-specific) and it handle well most common combinations of species names
and their authorities. However, the function has not been tested for a wide
range of possibilities, so some level of double-checking may be necessary.
}
\examples{

\dontrun{
nomes <- c("Lindsaea lancea", "lindsaea lancea", "lindsaea Lancea",
          "Lindsaea Lancea", "LINDSAEA LANCEA", "Lindsaea lancea var. Angulata",
          "Lindsaea lancea (L.) Bedd.", "Lindsaea Lancea (L.) Bedd.",
          "lindsaea lancea (L.) Bedd.",
          "Lindsaea", "Lindsaea Dryand. ex Sm.",
          "LINDSAEA", "LINDSAEA Dryand. ex Sm.", "LINDSAEA LANCEA (L.) Bedd.",
          "Lindsaea lancea var. angulata Rosenst.",
          "Lindsaea lancea angulata Rosenst.",
          "LINDSAEA LANCEA ANGULATA Rosenst.",
          "Blechnum antillanum Proctor",
          "Blechnum occidentale leopoldense Dutra",
          "Blechnum occidentale var. leopoldense Dutra")

fixCase(nomes)
}

}
\author{
Renato A. F. de Lima
}
\keyword{internal}
