% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixEncoding.R
\name{fixEncoding}
\alias{fixEncoding}
\title{Fix Encoding Problems}
\usage{
fixEncoding(x)
}
\arguments{
\item{x}{a vector with characters to be replaced}
}
\value{
the input vector with the encoding replaced
}
\description{
Simple function to replace typical encoding problems when
importing text with accents or special characters which is not in UTF-8
(i.e. latin1 encoding).
}
\details{
The function makes no encoding guessing. It just directly replaces
common enconding problems from text in latin1 to UTF-8 characters. So, make
sure that this replacement makes sure to your data.
}
\examples{
  nomes <- c("À", "Â", "Ù", "Ú")

\dontrun{
  fixEncoding(nomes)
}

}
\author{
Renato A. F. de Lima & Leila Meyer
}
\keyword{internal}
