% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixField.R
\name{fixField}
\alias{fixField}
\title{Fixing and Formatting the Herbarium Fields}
\usage{
fixField(x, origin = NULL)
}
\arguments{
\item{x}{a data frame}

\item{origin}{code of collection from which the input data was downloaded}
}
\value{
the data frame \code{x} in the proper format. It also returns the
required field names that are missing and those that were replaced or dropped.
}
\description{
Puts the occurrence data frame in the right format for data processing and validation
}
\details{
This function ...
}
\examples{
 ## Example currently not available ##


}
\author{
Lima, R.A.F.
}
