% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixLoc.R
\name{fixLoc}
\alias{fixLoc}
\title{Edit and Format Locality Information}
\usage{
fixLoc(x, adm.levels = c("country", "stateProvince", "municipality",
  "locality"), scrap = TRUE)
}
\arguments{
\item{x}{a data.frame}
}
\description{

}
