% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixName.R
\name{fixName}
\alias{fixName}
\title{Preparing People's Name}
\usage{
fixName(x, special.char = FALSE, from = "UTF-8",
  to = "windows-1252//TRANSLIT")
}
\arguments{
\item{x}{the character string.}
}
\value{
the character string \code{x} in the starndard needed for further processing.
}
\description{
Fix problems and stadardize name notation
}
\details{
The function fixes problems in name notation (e.g. orfan spaces) and
standardized the notation (i.e. Faria Jr. to Faria Junior). Due to encoding
problems, names are return without accents by default. But user can choose
between outputs with and without accents and species characters.
}
\examples{
  fixName("J.E.Q. Faria Jr.")
  fixName("J.E.Q. Faria Jr.", special.char = TRUE)
  fixName("Leitão F°, H.F.")
  fixName("Gert G. Hatschbach, et al.")
  fixName("Karl Emrich & Balduino Rambo")
  fixName('F. da S.N. Thom\\xe9')
  fixName('Pedro L.R.de Morães')
}
\author{
Lima, R.A.F.
}
