% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixSpecies.R
\name{fixSpecies}
\alias{fixSpecies}
\title{Edit Scientific Name Notation}
\usage{
fixSpecies(
  x = NULL,
  tax.name = "scientificName",
  author.name = "scientificNameAuthorship",
  rm.rank = FALSE,
  rm.indet = FALSE
)
}
\arguments{
\item{x}{a vector or data.frame containing the taxon name
information}

\item{tax.name}{character. The name of the column containing the
taxon name. Defaults to "scientificName"}

\item{author.name}{character. The name of the column containing the
author of scientific name. Defaults to "scientificNameAuthorship"}

\item{rm.rank}{logical. Should the infra-specific rank abbreviation
be removed from the name? Defaults to FALSE}

\item{rm.indet}{logical. Should the abbreviations for unspecific
names (i.e. sp. or spp.) be removed? Defaults to FALSE}
}
\value{
A data frame with the input vector or data frame plus new columns
\code{scientificName.new} (suggestion for a more correct taxon name),
\code{scientificNameAuthorship.new} (suggestion for the authorship of
the taxon name, if present) and \code{scientificNameStatus} (all the
flags detected in the original taxon name. See Details for a
description of all possible flags in the column
\code{scientificNameStatus}.
}
\description{
Identifies open nomenclature (aff., cf.) in scientific
names, classification under species level (var., subsp., f.) and
standardize the notation of incomplete name identifications. It
creates a new column with the new suggested name and divides
taxon names from their authors. It also flags and solve some
problematic names (character string with numbers, wrong case, or
other names besides genus and epithet etc). Names can be returned
with or without infra-specific ranks (var., subsp. and f.) or
abbreviations of unspecific names (sp. or spp.).
}
\details{
These function deals with most format and particular cases of taxon
names but of course it does not solve all of them. There are some
problems related to how the function deals with some hybrid species
notation and for some cases of genus or higher level taxonomy with
authors. See examples for those cases when the function does not
work as expected.

Possible flags returned in \code{scientificNameStatus}: \describe{
\item{\code{possibly_ok}}{scientific name following the classic pattern
'Genus epithet'}
\item{\code{not_Genus_epithet_format}}{scientific name not following
the expected pattern Genus epithet}
\item{\code{variety}}{scientific name with variety}
\item{\code{subspecies}}{scientific name with subspecies}
\item{\code{forma}}{scientific name with form}
\item{\code{infra_specific}}{scientific name with genus, specific epithet and
infra-specific, but no infra-specific rank}
\item{\code{hybrid}}{scientific name of a hybrid species}
\item{\code{conferre}}{open nomenclature cf. in the scientific name}
\item{\code{affinis}}{open nomenclature aff. in the scientific name}
\item{\code{indet}}{taxon identified only at genus level}
\item{\code{subfamily_as_genus}}{subfamily as genus, not a valid name}
\item{\code{family_as_genus}}{family as genus, not a valid name}
\item{\code{order_as_genus}}{order as genus, not a valid name}
\item{\code{incertae_sedis}}{scientific name of uncertain placement}
\item{\code{species_nova}}{species name contains an indication of a new
species, possibly not yet a valid name}
\item{\code{name_w_authors}}{scientific name has authors}
\item{\code{name_w_wrong_case}}{scientific name has upper/lowercase issues}
\item{\code{name_w_non_ascii}}{species name has non ASCII characters, not a
valid name}
\item{\code{abbreviated_genus}}{genus is abbreviated}
\item{\code{name_w_digits}}{scientific name has digits} }
}
\examples{

# Two last names are examples for which the function cannot handle yet.

df <- data.frame(scientificName =
c("Lindsaea lancea", "Lindsaea lancea (L.) Bedd.",
"Lindsaea lancea var. Angulata",
"Lindsaea Aff. lancea",
"Lindsaea", "Lindsaea sp.", "Lindsaeaceae sp.",
"Lindsaea aff. lancea (L.) Bedd.",
"Lindsaea ×improvisa K.U.Kramer",
"Parablechnum C.Presl",
"Blechnum spannagelii Rosenst.",
"Blechnum occidentale leopoldense Dutra",
"Blechnum austrobrasilianum de la Sota",
"Urbanodendron Mez",
"Arabidopsis thaliana × Arabidopsis arenosa"))

fixSpecies(df)
fixSpecies(df, rm.rank = TRUE)
fixSpecies(df, rm.rank = TRUE, rm.indet = TRUE)

}
\references{
Sigovini, M., Keppel, E. and Tagliapietra, D. (2016) Open
Nomenclature in the biodiversity era. Methods in Ecology and
Evolution 7(10): 1217-1225.
}
\seealso{
Functions \link[plantR]{fixAnnotation}, \link[plantR]{fixIndet},
\link[plantR]{fixCase} and \link[plantR]{fixAuthors}
}
\author{
Sara Mortara & Renato A. Ferreira de Lima
}
