% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatCoord.R
\name{formatCoord}
\alias{formatCoord}
\title{Format Geographical Coordinates}
\usage{
formatCoord(
  coords,
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  flag = TRUE,
  lat.orig = "decimalLatitude",
  lon.orig = "decimalLongitude",
  lat.gazet = "latitude.gazetteer",
  lon.gazet = "longitude.gazetteer",
  res.gazet = "resolution.gazetteer",
  lat.new = "decimalLatitude.new",
  lon.new = "decimalLongitude.new",
  rm.gazet = FALSE
)
}
\arguments{
\item{coords}{data.frame}

\item{lat}{character. The name of the column containing the latitude. Default
to "decimalLatitude".}

\item{lon}{character. The name of the column containing the longitude.
Default to "decimalLongitude".}

\item{flag}{logical. Should the differences between the original and edited
coordinates be flagged? Default to TRUE.}

\item{lat.orig}{character. The name of the column containing the latitude. Default
to "decimalLatitude".}

\item{lon.orig}{character. The name of the column containing the longitude.
Default to "decimalLongitude".}

\item{lat.gazet}{character. The name of the column containing the latitude
from the gazetteer. Default to "latitude.gazetteer".}

\item{lon.gazet}{character. The name of the column containing the longitude
from the gazetteer. Default to "longitude.gazetteer".}

\item{res.gazet}{character. The name of the column containing the resolution
of the locality information from which the gazetteer coordinates were
obtained.}

\item{lat.new}{character. The name of the column containing the edited
latitude, combining the coordinates from the source data and the gazetteer.
Default to "decimalLatitude.new".}

\item{lon.new}{character. The name of the column containing the working
longitude combining the coordinates from the source data and the gazetteer.
Default to "decimalLongitude.new".}

\item{rm.gazet}{logical. Should the coordinates form the gazetterr be removed
from the data after the replacement of missing coordinates? Default to FALSE.}
}
\value{
The input data frame \code{coords}, plus the new columns
'origin.coord' and 'precision.coord' with the formatted information. The
new columns have the same name as proposed by the Darwin Core standards
followed by the suffix '.new'.
}
\description{
This function formats geographical coordinates to decimal
degrees and replaces missing coordinates by the coordinates obtained from
the gazetteer.
}
\details{
The function works as a wrapper, where the individuals steps of the
proposed \strong{plantR} workflow for editing the geographical coordinates are
performed altogether (see the \strong{plantR} tutorial and the help of each
function for details).

The input data frame usually contains the following fields (the
function defaults):
\itemize{
\item the source coordinates: 'decimalLongitude' and 'decimalLatitude';
\item the gazetteer coordinates: 'latitude.gazetteer' and 'longitude.gazetteer';
\item the precision of the gazetteer coordinates: 'resolution.gazetteer';
\item the working coordinates: 'decimalLatitude.new' and 'decimalLongitude.new'.
}
}
\seealso{
\link[plantR]{prepCoord} and \link[plantR]{getCoord}.
}
