% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatFamily.R
\name{formatFamily}
\alias{formatFamily}
\title{Standardize Botanical Family}
\usage{
formatFamily(x)
}
\arguments{
\item{x}{a data frame containing at least the columns `family`, `genus` and
`speciesName`.}
}
\value{
the data frame \code{x} with an additional column called 'family.new'.
}
\description{
The functions use the ...
}
\details{
First search of names is based on the list of families and accepted synonyms
for vascular plants obtained from the APG website
(\href{http://www.mobot.org/MOBOT/research/APweb/}), which includes families
cited in the APG IV (2016) and in the PPG I (2016). If the family name is not
found in the APG list, a search is carried in the Brazilian Flora 2020
(BF-2020), using package `flora`. If the family name is still not found in
BF-2020, then a final try is performed in The Plant List (TPL), using package
`Taxonstand`. Family names retrieved from BF-2020 and TPL are finally
converted to the names accepted by the APG IV or PPG I.

In case there is a conflict in the original family name and the name found
based on the genus name, the original name is replaced by the name from the
APG IV or PPG I with a warning.
}
\examples{

occs <- data.frame(family = c("Ulmaceae", "Cannabaceae", "Salicaceae",
"Flacourtiaceae", "Vivianiaceae"), genus = c("Trema", "Trema", "Casearia",
"Casearia", "Casearia"), speciesName = c("Trema micrantha", "Trema
micrantha", "Casearia sylvestris","Casearia sylvestris","Casearia
sylvestris"), stringsAsFactor = FALSE)
formatFamily(occs)

}
\references{
Angiosperm Phylogeny Group (2016). An update of the Angiosperm
Phylogeny Group classification for the orders and families of flowering
plants: APG IV. Bot. J. Linnean Soc. 181: 1-20.

PPG I (2016). A community-derived classification for extant lycophytes and
ferns. Journal of Systematics and Evolution. 54 (6): 563–603.
}
\author{
Renato A. F. de Lima
}
