% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatOcc.R
\name{formatOcc}
\alias{formatOcc}
\title{Format Names, Numbers and Dates}
\usage{
formatOcc(x, noNumb, noYear, noName)
}
\arguments{
\item{x}{a data frame or data table}

\item{noNumb}{standard for missing data in Number}

\item{noYear}{standard for missing data in Number}

\item{noName}{standard for missing data in Number}
}
\description{
This function standardizes names, number and dates from data obtained from on-line data bases such as GBIF and speciesLink
}
