% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTax.R
\name{formatTax}
\alias{formatTax}
\title{Format Taxonomic Information}
\usage{
formatTax(
  tax,
  use.suggestion = TRUE,
  tax.name = "scientificName",
  author.name = "scientificNameAuthorship",
  rm.rank = FALSE,
  tax.names = c("scientificName.new", "scientificNameAuthorship.new"),
  db = "bfo",
  sug.dist = 0.9,
  use.authors = TRUE,
  clean.names = FALSE,
  split.letters = FALSE,
  parallel = FALSE,
  cores = 2,
  drop.cols = c("match_type", "multiple_match", "fuzzy_dist_name", "fuzzy_dist_author",
    "name.status", "taxon.status", "accepted.name", "accepted.authorship",
    "accepted.taxon.rank", "accepted.name.status"),
  fam.name = "family",
  gen.name = "genus",
  spp.name = "scientificName"
)
}
\arguments{
\item{tax}{a data.frame containing the taxonomic information to be edited.}

\item{use.suggestion}{logical. Should the edited species names be used
instead of the input species names? Defaults to TRUE.}

\item{tax.name}{character. The name of the column containing the
taxon name. Defaults to "scientificName"}

\item{author.name}{character. The name of the column containing the
author of scientific name. Defaults to "scientificNameAuthorship"}

\item{rm.rank}{logical. Should the infra-specific rank abbreviation
be removed from the name? Defaults to FALSE}

\item{tax.names}{character. Names of the columns containing the
taxon names and authorships. Defaults to 'scientificName.new' and
'scientificNameAuthorship.new'.}

\item{db}{the taxonomic backbones to be consulted for valid names,
in their preferred order of priority. By default, only the
Brazilian Flora 2020 taxonomic backbone ('bfo') is used.}

\item{sug.dist}{a fraction expressing the maximum distance allowed
between the taxon name and the suggested names in the backbone,
which is passed to the argument \code{max.dist} of the function
\code{nameMatching()}. Defaults to 0.9 (i.e. maximum of 10\%
difference).}

\item{use.authors}{logical. Should authorships be used in the name
matching process? Default to TRUE.}

\item{clean.names}{logical. Should spaces, ponctuation, symbols and
species characters be removed prior to name matching? Defaults to
TRUE.}

\item{split.letters}{logical. Should fuzzy matching be performed
separately by the input and reference name initials? Defaults to
FALSE}

\item{parallel}{logical. Should fuzzy matching be performed in
parallel? Defaults to FALSE}

\item{cores}{integer. The number of cores for parallel execution.
Two by default.}

\item{drop.cols}{character. Name of columns from the reference
backbone that should be dropped from the results.}

\item{fam.name}{character. The name of the column containing the family
names. Default to "family".}

\item{gen.name}{character. The name of the column containing the genus names.
Default to "genus"}

\item{spp.name}{character. The name of the column containing the species
names. Default to "scientificName"}
}
\value{
The input data frame \code{tax}, plus the new columns with the
formatted information. The new columns have the same name of the
Darwin Core standards, but followed by the suffix '.new'.
}
\description{
This function edits and standardizes the names of plant
species and families
}
\details{
The function works as a wrapper, where the individuals steps of the
proposed \strong{plantR} workflow for editing taxonomic information are
performed altogether (see the \strong{plantR} tutorial and the help of each
function for details).

The input data frame usually contains the following taxonomic fields:
"family", "genus", "scientificName" and "scientificNameAuthorship". But
users can define the names for their own data.
}
\seealso{
\link[plantR]{fixSpecies}, \link[plantR]{prepSpecies} and
\link[plantR]{prepFamily}.
}
