% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdmin.R
\name{getAdmin}
\alias{getAdmin}
\title{Get Administrative Subdivisions from a Gazetteer}
\usage{
getAdmin(
  x,
  gazet = "plantR",
  str.name = "loc.correct",
  gazet.names = c("loc.correct", "NAME_0", "NAME_1", "NAME_2", "NAME_3", "source")
)
}
\arguments{
\item{x}{a vector or a dataframe containing the \strong{plantR} search string
in the proper format (lowercase, separated by an underline).}

\item{gazet}{a data frame containing the gazetteer with the list of
names and locality subdivisions. The default is "plantR", the internal
\strong{plantR} gazetteer (focused on Latin America and the Caribbean).}

\item{str.name}{a character corresponding to the columns name containing the
locality search string to match the input dataframe and the gazetteer.
Defaults to 'loc.correct'.}

\item{gazet.names}{a vector of the column names containing the locality
search string in the gazetteer, and the administrative subdivisions, in
that order. If available, the source of the administrative information can
be provided as the last name of this vector. Defaults to columns names of
the \strong{plantR} gazetteer: 'loc.correct', 'NAME_0', 'NAME_1', 'NAME_2',
'NAME_3 and 'source'.}
}
\value{
A data frame containing the input locality string in the first
column, the administrative names at the resolution retrieved (four
following columns) and the source of the administrative names (last
column)
}
\description{
Obtain the names of the administrative subdivisions for a given
locality from a \href{https://en.wikipedia.org/wiki/Gazetteer}{gazetteer}.
Users can use the default \strong{plantR} gazetteer or provide one to the
function.
}
\details{
Although the internal gazetteer used in \strong{plantR} can handle common
notation variants and typos in locality names (see function \code{getLoc}), the
function \code{getAdmin} assumes that these issues were already solved.
Otherwise, the function may not find the information required. See Examples
below.

The retrieval of information depends on the completeness of the gazetteer
itself. So, even if all variants and typos were solved for an existing
locality, if the gazetteer does not include a specific locality, the query
will return NAs. The gazetteer is permanently being improved. If you found
an error or want to contribute with region-specific gazetteers, please send
an email to \href{mailto:renato.lima@naturalis.nl}{renato.lima@naturalis.nl}.

In Brazil, NAME_0, NAME_1, NAME_2, AND NAME_3 correspond to the Country,
State, Municipality and Locality, respectively. But this may not be the
case of other countries if they adopt a different nomenclature for their
administrative levels.

Finally, function \code{getAdmin} currently does not return information below
the municipality level, although the default \strong{plantR} gazetteer can retrieve
information at lower administrative levels (i.e. locality and sub-locality).

A different gazetteer than the default one can be provided using the
argument \code{gazet}. If the column names of the user-provided gazetteer do not
match those of the default gazetteer, the column names can be supplied
using argument \code{gazet.names}.
}
\examples{
str <- c("paraguay_paraguari",
         "brazil_parana_paranagua",
         "brazil_rio janeiro_parati", # an example of a variant in the locality name
         "brazil_rio janeiro_paraty",
         "brazil_sao paulo_sao miguel arcanjo_pe carlos botelho",
         "united states_florida") # valid location but not in the default gazetteer
getAdmin(str)

}
\author{
Renato A. F. de Lima
}
