% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCult.R
\name{getCult}
\alias{getCult}
\title{Flag Occurrences From Cultivated Individuals}
\usage{
getCult(x, cc_inst = FALSE)
}
\arguments{
\item{x}{a data frame with the occurrence data.}

\item{cc_inst}{logical. Should the identification of occurrences in the
vicinity of biodiversity institutions (package `coordinateCleaner`) be performed?
Default to FALSE.}
}
\description{
This function ...
}
\details{
The input data frame \code{x} must contain at least the columns
  containing the description of the occurrence (i.e. 'occurrenceRemarks') or
  the locality description (e.g. 'locality').
}
