% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDup.R
\name{getDup}
\alias{getDup}
\title{Search For Duplicate Specimens Among Collections}
\usage{
getDup(df, flag.ind = TRUE, ...)
}
\arguments{
\item{df}{a data frame with the colection code (first column) and the strings
to be used for the duplicate search (second and other columns), typically
the exit of the function `prepDup`.}

\item{flag.ind}{logical. Should duplicates based only on indirect duplicates
be flagged with brackets? Default to TRUE.}
}
\value{
see: https://stackoverflow.com/questions/56740990/optimizing-grepl-and-other-operations-inside-a-loop-using-data-table
}
\description{
This function ...
}
\details{
The input data frame \code{df} must contain at least the columns ...
}
\author{
Renato A. F. de Lima
}
