% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFamily.R
\name{getFamily}
\alias{getFamily}
\title{Get Family Name}
\usage{
getFamily(x = NULL, db = "bfo", fuzzy.match = FALSE)
}
\arguments{
\item{x}{a vector of genus names with no default.}

\item{db}{the taxonomic backbone to be consulted for family names.
By default, the Brazilian Flora 2020 taxonomic backbone
('bfo') is used.}

\item{fuzzy.match}{logical. Whether if fuzzy matching on genus names
should be conducted. Defaults to FALSE (no fuzzy matching)}
}
\value{
a vector of the same length as \code{x} with the family
names matched in the taxonomic backbone
}
\description{
Search for family names based on a
vector of genus names and a taxonomic backbone
}
\details{
The exact and fuzzy matching (if selected) between \code{x} and the
names of genera present in the backbone (argument \code{db}) are only
performed based on currently accepted genus names in the taxonomic
backbone (i.e. backbone column taxon.status equal to 'accepted').

Currently, fuzzy macthes are conducted using the defaults of
function \code{nameMatching()}, which is used internally.
}
\examples{

genus <- c("Trema", "Casearia", "Turnera", "Viviania", "Dombeya",
        "Casearea", "Indet.", "",  NA)
getFamily(genus)
getFamily(genus, fuzzy.match = TRUE)

}
\author{
Renato A. Ferreira de Lima
}
