% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLoc.R
\name{getLoc}
\alias{getLoc}
\title{Get Locality and Coordinates}
\usage{
getLoc(x, gazet = "plantR")
}
\arguments{
\item{x}{data frame with the columns: "country.new","stateProvince.new","municipality.new","locality.new"}
}
\description{
Create the locality string used to search for coordinates in a gazetteer. This information is
used to obtain missing coordinates based on locality descriptions and in the process of validation of the
geographical coordinates.
}
