% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastName.R
\name{lastName}
\alias{lastName}
\title{Get Last Name}
\usage{
lastName(name, noName = "s.n.")
}
\arguments{
\item{name}{the character string or vector containing the names.}

\item{noName}{character. The standard notation for missing names in
\code{name}. Default to "s.n.".}
}
\value{
the last name provided in \code{name} with the first letter
  capitalized.
}
\description{
Extract the last name of the collector or determiner of a
  biological specimen.
}
\details{
The function works for simple last names or compound last names,
  independently if names are provided in lower or capital letters. It is
  relatively stable to the format and spacing of the string provided, but it
  may not work in all cases.

  It implicitly assumes that last names are the ones at the end of the name or
  the first name preceeding the comma.

  If only one name is given, the function returns the same name with the first
  letter capitalized.

  For missing names (i.e. "", " " or NAs) the function returns the character
  defined by the argument `noName`
}
\examples{

# Simple last name
 lastName("Al Gentry")
 lastName("Gert Guenther Hatschbach")
 lastName("Gert G. Hatschbach")
 lastName("GERT GUENTHER HATSCHBACH")
 lastName("HATSCHBACH, G.G.")
 lastName("HATSCHBACH, G. G.")
 lastName("G. G. HATSCHBACH")

 # Last name with generational suffixes
 lastName("Hermogenes Leitao Filho")
 lastName("Leitao Filho, H.")
 lastName("Leitao Filho, H.F.")
 lastName("Filho Neto, S.J.")

 # Compound last name
 lastName("Augustin Saint-hilaire")
 lastName("Saint-Hilaire A.")

 # Multiple names
 names <- c("Gentry, AH", "Gentry A.H.",
 "Leitao filho, H. F.","Saint-Hilaire, augustin")
 lastName(names)

 # Unusual formatting
 lastName("Cesar Sandro, Esteves, F")
 lastName("Mendonca Filho, C.V. Neto, F.C.C.")
 # two or more names, separated by comma: output correct
 lastName("A. Alvarez, A. Zamora & V. Huaraca")
 # two or more names, not separated by comma: output incorrect (give names of authors)
 lastName("Karl Emrich & Balduino Rambo")

}
\author{
Renato A. F. de Lima & Hans ter Steege
}
