% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latamMap}
\alias{latamMap}
\title{Maps of Latin American countries}
\format{A list of 44 Multipolygon 'sf' objects}
\source{
\url{https://gadm.org}
}
\usage{
data(latamMap)
}
\description{
Maps used to perform the validation of the original geographical coordinates.
For each country, the map constains the lowest administrative level
available. Country, state and county (if present) names were edited to match
the `plantR` default gazetteer and original polygons were simplifyied
(tolerance 0.0001 decimal degrees). Reference system: WSG84
}
\keyword{datasets}
