% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missName.R
\name{missName}
\alias{missName}
\title{Homogenize Notation For Missing Name}
\usage{
missName(x, type = NULL, noName = "Anonymous")
}
\arguments{
\item{x}{the character string}

\item{type}{type of ...}

\item{noName}{standard for missing data in Name}
}
\value{
the character string \code{x} in the fixed TDWG format.
}
\description{
Standardize the different notation of missing collector or identifier names
}
\details{

}
\examples{
  missName(c("Gentry, AH", "s/col.", NA, "?", "s/c", "s/coletor"), type="collector", noName = "s./c.")
  missName(c("Gentry, AH", "s/col.", NA, "?", "s/c", "s/coletor"), type="collector", noName = "Anonymous")
  missName(c("Gentry, AH", "s/det.","s/det","s/d","Determiner unknown"), type="identificator", noName = "Anonymous")
}
\references{

}
\author{
Lima, R.A.F.
}
