% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepCoord.R
\name{prepCoord}
\alias{prepCoord}
\title{Format Geographical Coordinates}
\usage{
prepCoord(x)
}
\arguments{
\item{x}{a data.frame containing the geographical coordinates.}
}
\value{
The data frame \code{x} with the edited geographical coordinates, in
  decimal degrees.
}
\description{
Format geographical coordinates do decimal degrees.
}
\details{
This is a simple function that works for standardizing common issues
 in geographical coordinates, such as format and spacing. It was designed to
 process coordinates already in decimal degrees, so, it does not handle all
 formats and may return `NAs` even if a coordinates is given and valid.

Currently, the function does not differentiate missing coordinates from those
 provided in a format that `prepCoord` does not handle (e.g. UTM). It also
 does not differentiate original coordinates in decimal degrees in a good
 format from those that needed to be formatted (i.e. no flagging is returned).
 Function should work fine for in the degress + decimal minutes format. See
 examples below.
}
\examples{

## Creating some geographical coordinates in different formats
(coords <- data.frame(decimalLatitude = c(-23.475389, "-23,475389", "23o 28' 31.40\"S", -93.475389,
                                          0, "missing", "blocked", NA, "23° 28.5233'S","283088.52 E"),
                      decimalLongitude = c(-47.123768, "-47,123768", "47o 07' 25.56\"W", -47.123768,
                                          0, "missing", "blocked", NA, "47° 07.4260'W","7402251.30 S")))

## Formatting the geographical coordinates
prepCoord(coords)

}
\author{
Renato A. F. de Lima
}
