% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepLoc.R
\name{prepLoc}
\alias{prepLoc}
\title{Format Locality String}
\usage{
prepLoc(x)
}
\arguments{
\item{x}{character. The concatenated locality string to be simplified.}
}
\value{
The edited locality string.
}
\description{
Simplify the notation of the standard \strong{plantR} locality
string, which is used for the validate localities and geographical
coordinates.
}
\details{
The goal of this function is to simplify the locality strings in
order to reduce variations in locality notation and thus to increase the
chances of match with the \strong{plantR} gazetteer and maps.

It tranforms the string to lowercase and removes special characters and
prepositions of locality names.
}
\examples{
  prepLoc("Brazil_Rio de Janeiro")
  prepLoc("Brazil_Rio Grande do Sul_Coqueiros do Sul")
  prepLoc("Brazil_Bahia_Arraial d'Ajuda")
  prepLoc("Brazil_S\xE3o Paulo_Sete Barras_Parque Estadual de Carlos Botelho")
  prepLoc("Cuba_Isla de la Juventud")

}
\seealso{
\link[plantR]{strLoc}
}
\author{
Renato A. F. de Lima
}
