% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepLoc.R
\name{prepLoc}
\alias{prepLoc}
\title{Format Locality String}
\usage{
prepLoc(x)
}
\arguments{
\item{x}{character. The concatenated locality string to be simplified.}
}
\value{
The edited locality strings.
}
\description{
Simplify the notation of the `plantR` string of locality search
}
\details{
The function removes unwanted hifens and prepositions of locality names
}
\examples{
  prepLoc("Brazil_Rio de Janeiro")
  prepLoc("Brazil_Rio Grande do Sul_Coqueiros do Sul")
  prepLoc("Brazil_Bahia_Arraial d'Ajuda")
  prepLoc("Brazil_São Paulo_Sete Barras_Parque Estadual de Carlos Botelho")

}
\author{
Renato A. F. de Lima
}
