% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdwgName_new.R
\name{prepName}
\alias{prepName}
\title{Format People's Name To TDWG Standard}
\usage{
prepName(x, fix.names = TRUE, prep.tdwg = TRUE)
}
\arguments{
\item{x}{the character string.}

\item{fix.names}{logical. Should the general notation of names be
standardized? Default to TRUE.}

\item{prep.tdwg}{logical. Should the names already in the 'Lastname,
Name/Initials' format be standardized? Default to TRUE.}

\item{...}{Parameters from fixName}
}
\value{
The character string \code{x} in the TDWG format.
}
\description{
Format or convert multiple collector or identificator names into the
  \href{https://www.tdwg.org/}{Biodiversity Information Standards} (TDWG)
  format.
}
\details{
The standard name format suggested by the
  \href{https://www.tdwg.org/}{TDWG} is: Last name, followed by a comma and
  then the initials, separated by points (e.g. Hatschbach, G.G.).

  The function identifies (and removes) name prefixes or prepositions (e.g. de,
  dos, van, ter, ...). Also, it removes some titles (i.e. Dr., Prof., Pe.)
  but not all of them (e.g. Doctor, Priest, etc.). The function also does not
  handle hyphenated first names. If only one name is given, the function
  returns \code{x} with the first letter capitalized.

  The function has the option of stardardizing the general notation of names
  and the general format of names already in the 'Lastname, Name/Initials'
  format. This standardization is controled by the arguments `fix.names` and
  `prep.tdwg`, which call internally the __plantR__ functions `fixName()`
  and `prepTDWG()`. So, the function output it is relatively stable regarding
  the input format and spacing, but it may not work in all cases,
  particularly if strings provided already contain commas.
}
\examples{
  # Simple name
  prepName("Al Gentry")
  prepName("Alwyn Howard Gentry")
  prepName("Gert G. Hatschbach")
  prepName("HATSCHBACH, G.G.")
  prepName("HATSCHBACH, G. G.")
  prepName("G. G. HATSCHBACH")

  # Name with prepositions
  prepName("Carl Friedrich Philipp von Martius")
  prepName("Alphonse Louis Pierre Pyrame de Candolle")
  prepName("Simon Jan van Ooststroom")
  prepName("Jan van der Hoeven") ##CHECK!
  prepName("Maria da Silva")
  prepName("Silva, M. da")
  prepName("da Silva, M.")

  # Name with generational suffixes
  prepName("Hermogenes Leitao Filho")
  prepName("Leitao Filho, H.")
  prepName("Leitao Filho, H.F.")
  prepName("Leitao Filho, HF") ##CHECK!
  prepName("S.J. Filho Neto")

  # Compound last name (needs to be marked with a '-')
  prepName("Augustin Saint-hilaire")
  prepName("Saint-Hilaire A.")
  prepName("Augustin Saint Hilaire") # compound name missing '-'

  # Names with titles
  prepName("Pe. Raulino Reitz")
  prepName("Prof. Hermogenes de Freitas Leitao Filho")

  # Unusual formatting (function won't always work)
  prepName("[D. Hugh-Jones]") #names inside bracket: output correct
  prepName("Cyl Farney Catarino de Sa") # small last name, no comma: output correct
  prepName("Cyl Sa") # small last name, without comma: output incorrect (inverted)
  prepName("Sa, Cyl") # small last name, with comma: output incorrect (inverted)
  prepName("L. McDade") #cannot recognize names starting with Mc or Mac
  prepName("Cesar Sandro, Esteves, F") # one name, two commas: fails to get the right last name
  # two or more names: output incorrect (combine names of authors)
  prepName("Mendonca Filho, C.V.; Neto, F.C.C.")
  # two or more names, separeted by comma: output incorrect (combine names of authors)
  prepName("A. Alvarez, A. Zamora & V. Huaraca")
  # two names, not separated by comma: output incorrect (combine names of authors)
  prepName("Karl Emrich & Balduino Rambo")

}
\references{
Conn, Barry J. (ed.) (1996). HISPID 3 - Herbarium Information Standards and
  Protocols for Interchange of Data. Herbarium Information Systems Committee'
  (HISCOM). https://www.tdwg.org/standards/hispid3/

Willemse, L.P., van Welzen, P.C. & Mols, J.B. (2008).
  Standardisation in data-entry across databases: Avoiding Babylonian
  confusion. Taxon 57(2): 343-345.
}
\seealso{
\link[plantR]{fixName}, \link[plantR]{prepTDWG}, \link[plantR]{lastName} and
 \link[plantR]{getTDWG}.
}
\author{
Renato A. F. de Lima & Hans ter Steege
}
