% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSpecies.R
\name{prepSpecies}
\alias{prepSpecies}
\title{Get Valid Taxon Names}
\usage{
prepSpecies(
  x,
  tax.names = c("scientificName", "scientificNameAuthorship"),
  db = "bfo",
  use.authors = TRUE,
  replace.names = TRUE,
  mult.matches = "all",
  sug.dist = 0.9,
  clean.indet = TRUE,
  clean.names = FALSE,
  dist.method = "jw",
  split.letters = FALSE,
  parallel = FALSE,
  cores = 2,
  show.progress = FALSE,
  auth.factor = 0.5,
  auth.factor.max = 3,
  drop.cols = c("match_type", "multiple_match", "fuzzy_dist_name", "fuzzy_dist_author",
    "name.status", "taxon.status", "accepted.id", "accepted.tax.name",
    "accepted.tax.authorship", "accepted.taxon.rank", "accepted.taxon.status",
    "accepted.name.status", "taxon.distribution")
)
}
\arguments{
\item{x}{a vector or data.frame of taxon names and their
authorships, ideally the output from function \code{fixSpecies()}.}

\item{tax.names}{character. Names of the columns containing the
taxon names and authorships. Defaults to 'scientificName' and
'scientificNameAuthorship'.}

\item{db}{the taxonomic backbones to be consulted for valid names,
in their preferred order of priority. By default, only the
Brazilian Flora 2020 taxonomic backbone ('bfo') is used.}

\item{use.authors}{logical. Should authorships be used in the name
matching process? Default to TRUE.}

\item{replace.names}{logical. If taxon names should be replaced
based on the results of the name checking (e.g. if synonyms
should be replaced by its accepted name. Default to TRUE.}

\item{mult.matches}{a character declaring how to deal with multiple
name matches. Options are 'all' (concatenate all matches in the
same line) and 'best' (only one name match is returned per line).
Defaults to 'all'.}

\item{sug.dist}{a fraction expressing the maximum distance allowed
between the taxon name and the suggested names in the backbone,
which is passed to the argument \code{max.dist} of the function
\code{nameMatching()}. Defaults to 0.9 (i.e. maximum of 10\%
difference).}

\item{clean.indet}{logical. Should the markers and identifiers of
unidentified species (e.g. sp. or spp.) be removed prior to name
matching? Defaults to TRUE.}

\item{clean.names}{logical. Should spaces, punctuation, symbols and
species characters be removed prior to name matching? Defaults to
FALSE.}

\item{dist.method}{fuzzy matching algorithm to be passed on to
\code{stringdist::amatch()}. Defaults to "jw" (i.e. Jaro-Winkler)}

\item{split.letters}{logical. Should fuzzy matching be performed
separately by the input and reference name initials? Defaults to
FALSE}

\item{parallel}{logical. Should fuzzy matching be performed in
parallel? Defaults to FALSE}

\item{cores}{integer. The number of cores for parallel execution.
Two by default.}

\item{show.progress}{logical. Whether fuzzy matching progress
should displayed. Defaults to FALSE}

\item{auth.factor}{numeric. A proportion of extra increase in the
distance allowed for author names. By default, this value is 0.5,
which means that the distance for author names by default is
0.15, obtained using: (1 - sug.dist) + (1 - sug.dist) *
auth.factor. This distance is set to 1 with a warning if the
resulting value of the formula is higher than 1.}

\item{auth.factor.max}{numeric. Any factor to get a maximum
distance allowed for author names, from which any author matches
will be flagged as a bad match. By default, this factor is 3,
which means that the maximum distance for author names by default
is 0.45, obtained using: auth.factor.max *((1 - sug.dist) + (1 -
sug.dist) * auth.factor). This distance is set to 1 with a
warning if the resulting value of the formula is higher than 1.}

\item{drop.cols}{character. Name of columns from the reference
backbone that should be dropped from the results.}
}
\value{
A data frame with the same columns provided by the user and, by
default, some additional columns containing the suggested taxon
name ('suggestedName'), taxon authorship ('suggestedAuthorship'),
the corresponding scientific name ('scientificNameFull'), the
taxonomic information related to the suggested name (columns
'taxon.rank', tax.notes' and its 'id' in the reference taxonomic
backbone). The return of other columns can be controlled using the
argument \code{drop.cols}.
}
\description{
Validates taxon names (including the correction of
typos and the detection of orthographic variants) based on a
reference list of taxon names (e.g. taxonomic backbone) provided
by the user, \strong{plantR} (the default) or those obtained from the
companion package \strong{plantRdata}.
}
\details{
First, the function checks taxon names with authorship (if
available), using both exact and fuzzy matching. This is a good
practice to avoid multiple matches in the reference taxonomic
backbone related to homonyms. If a name below the threshold
distance defined in argument \code{sug.dist} is not found, then the
function does a second round of exact and fuzzy matching using
only the taxon names without authorship. This second check is
done for names provided without authorship by the user and
for names provided with authorship but very distant from
the one found in the backbone (i.e. bad fuzzy matches of
authorship names).

Distances between input and matched names and authorships are
computed without the hybrid symbol and without spaces,
respectively. These distances are computed separately and after
name matching, meaning that there is the possibility of fuzzy
matches with zero distances between input and reference names.
This is done to avoid inflating the number of names below the
threshold distances simply due to differences in format between
the input and reference names. This is particularly true for
author names, which can be quite variable in format between sources.

Defining reasonable distance values to flag bad fuzzy matched for
author names is quite arbitrary. In \code{prepSpecies()}, this
definition is done by combining the arguments \code{auth.factor} and
\code{auth.factor.max}, whose defaults were set based on a try and
error process using test names.

The name check compares all unique names in \code{x} with all names in
the taxonomic backbones selected. So, the speed of the checking
process depends on the number of name combinations in \code{x} and in
the backbone, besides the computer's operating system and
specifications. But if the backbone selected has many names (over
hundreds of thousands of names), such as those containing all
plant names for the entire world, the process may take up to a
minute to check a dozen of names.

But if you have many thousands of input names to check and a
large backbone, the function offers the possibility of doing the
name matching by initial letters and/or using parallelization.
See the help of function \code{nameMatching()} which is used internally.

The result of the name check will vary depending on the
completeness of the backbone. And do expect some false positive
matches, e.g. input names returning as misspelled but that
actually are not in the reference backbone (e.g. matching of
exotic names against a local or regional reference backbone). If
this is the case, one should use more restrictive values of
\code{sug.dist} than the default of 0.9. See examples for an
illustration of these differences.

Currently, there are two options: (1) using the \strong{plantR} internal
backbone for the Brazilian Vascular Flora (the default) or (2)
using user-provided taxonomic backbones. The choice between these
options is declared via the argument \code{db}, which is passed on to
the function \code{getTaxBackbone()}.

The companion R package \strong{plantRdata} provides different global
and regional taxonomic backbones in the exact format expected by
\code{prepSpecies()}. See the help of function \code{getTaxBackbone()} for
more details on this expected format.

Previous versions of this function used packages \strong{flora} and
\strong{Taxonstand} to provide suggested names. These packages are no
longer used, but the function default is to compare names against
the Brazilian Flora Online, the same backbones used by the
package \strong{flora}.

The function does not provide the suggested names from the
different backbones in separate columns. It combines into
the same column ('suggestedName') the suggestions from one or
more backbone. If more than one database is used, the suggested
names are combined following the same order of the database codes
provided in the argument \code{db}.

If \code{replace.names} is TRUE, the function replaces synonyms with
its accepted name and it changes the category 'synonym' in the
new column 'tax.notes' to 'replaced synonym'. When there is a bad
match, the dubious accepted name found is replaced by NA. In the
cases where multiple name matches have the same accepted name,
the accepted name for both names is also returned. In naddition,
when the backbone states that a name is a synonym but it does not
provide an accepted name, the name of the synonym is return and
the category on 'tax.notes' is changed to 'synonym not replaced'.

The function returns a new column named 'tax.notes' which
contains a summary of the result of the name matching for each
taxon name. See the details of the function \code{getTaxNotes} for
more details on those categories.

By default, the argument \code{mult.matches} returns all possible
names matched in the taxonomic backbone (mult.matches = 'all'),
which is especially important if the authorship is not provided.
If \code{mult.matches} is 'best', only one name is returned. Ideally,
this is the accepted name within the multiple matches in the
backbone. Result is a lot cleaner but in many cases this does not
means that the accepted name returned actually is the right
accepted name for the input name. So, users should be extra
careful when setting \code{mult.matches} to 'best'.

The output of this function contains columns which are reserved
within the \strong{plantR} workflow. These columns cannot be present in
the input data frame. The full list of reserved columns is stored
in the internal object \code{reservedColNames}.
}
\examples{
sp.list <- c("Casearia sylvestris","Casearia sylvestris",
  "Casearia sylvestirs", "Casearia silvestris",
  "Casearia sylvestris var. angustifolia", "Casearia attenuata",
  "Casearia celtidifolia","Casearia celtidifolia","Casearia celtidifolia",
  "Casearia serrulata", "Casearia serrulata", "Casearia serrulata",
  "Xylosma ciliatifolium", "Casearia tropicana")
aut_list <- c("Sw.", "", "Sw.", "", "Uittien", "Rusby",
  "", "Kunth", "Poepp. Eichler", "", "J. Seber ex Griseb.", "Sw.",
  "(Clos) Eichler", "L.")

df <- data.frame(scientificName = sp.list,
  scientificNameAuthorship = aut_list)

prepSpecies(df)
prepSpecies(df, sug.dist = 0.925)
prepSpecies(df, mult.matches = "best")


}
\seealso{
Functions \link[plantR]{getTaxBackbone},
\link[plantR]{nameMatching} and \link[plantR]{getTaxNotes}.
}
\author{
Renato A. Ferreira de Lima
}
