% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{Read Data From DwC-A Files}
\usage{
readData(
  file = NULL,
  path = "",
  dir.name = "",
  dir.tmp = "plantR_input",
  method = "auto",
  bind.data = TRUE,
  output = c("occurrence", "verbatim", "citations"),
  save = FALSE,
  file.format = "csv",
  compress = TRUE
)
}
\arguments{
\item{file}{character. Name of the DwC-A file (often a '.zip') containing the
species records.}

\item{path}{character. The path to the directory where the file was saved or
of the web Uniform Resource Locator (URL) from which the file can be
downloaded from. Default to the user working directory or to gbif download
path.}

\item{dir.name}{character. Name of the folder where the processed data should
be saved. Default the directory defined by \code{path}.}

\item{dir.tmp}{character. Name of the sub-folder where the temporary files
should be saved within \code{dir.name}. Default to "plantR_input".}

\item{method}{the method to be passed to function \code{download.file()} for
downloading files. Default to 'auto'.}

\item{bind.data}{logical. Should the occurrence and verbatim information be
combined into a single table? Default to TRUE.}

\item{output}{character. Which information from the Darwin-Core file should
be returned/saved? Default to 'occurrence', 'verbatim' and 'citations'.}

\item{save}{logical. Should the information be saved to file? Default to
FALSE.}

\item{file.format}{character. The file extension to be used for saving.
Default to 'csv'.}

\item{compress}{logical. Should the files be compressed? Default to TRUE.}
}
\description{
This function reads species records from Darwin Core Archive
(DwC-A) files, typically obtained from GBIF as a '.zip' file. It returns
different information from inside the DwC-A files so it can enter the
\strong{plantR} workflow. Optionally, it can be used to save the required
information into a local directory.
}
\details{
This function provides different options to read DwC-A files,
typically the ones obtained from GBIF. Currently, this zip file can be read
from a local directory or directly from GBIF API address. If the path is an
URL address (e.g. https://api.gbif.org/v1/occurrence/download/request),
then the function will download the zip file directly from the GBIF API.

The argument \code{output} defines which of the information within GBIF DwC-A
files should be returned. Currently, the outputs available are:
'occurrence', 'verbatim', 'multimedia', 'citations' and 'rights'. If more
than one output is selected, the function returns a list in which each
element represent the selected outputs. Currently, no data and database
metadata are returned (i.e. '.xml' files). See package \strong{finch} for the
complete parsing of DwC-A files and metadata.

All temporary files and folders are deleted after the extraction of the
information, except if \code{save} is TRUE. In this case, only the unzipped
files within the DwC-A file are removed.

Downloading large files (more than 2GB) may be an issue for some R
versions. The \code{method} 'wget' may be more appropriate for users with proxy
firewalls (see help of function \code{download.file()}).
}
\examples{
\dontrun{
  occs <- readData(file = "0227351-200613084148143.zip",
                   path <- "https://api.gbif.org/v1/occurrence/download/request/")
}

}
