% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspeciesLink.R
\name{rspeciesLink}
\alias{rspeciesLink}
\title{Gets occurrence data from speciesLink}
\usage{
rspeciesLink(
  dir = "results/",
  filename = "output",
  save = FALSE,
  key = NULL,
  basisOfRecord = NULL,
  family = NULL,
  species = NULL,
  collectionCode = NULL,
  country = NULL,
  stateProvince = NULL,
  county = NULL,
  Coordinates = NULL,
  CoordinatesQuality = NULL,
  Scope = NULL,
  Synonyms = "no synomyms",
  Typus = FALSE,
  Images = FALSE,
  RedList = NULL,
  MaxRecords = 5000,
  file.format = "csv",
  compress = FALSE
)
}
\arguments{
\item{dir}{Path to directory where the file will be saved. Default
is to create a "results/" directory}

\item{filename}{Name of the output file}

\item{save}{Logical. Save output to filename? Defaults to FALSE}

\item{key}{Character. Your personal specieslink API key.}

\item{basisOfRecord}{Character. Any in 'PreservedSpecimen',
'LivingSpecimen', 'FossilSpecimen', 'HumanObservation',
'MachineObservation' or 'MaterialSample'. Default is
'PreservedSpecimen' for museum and herbarium records.}

\item{family}{Character. The family name. More than one name should
be concatenated in a vector.}

\item{species}{Character. The genus name or the genus and epithet separated by
a space. More than one species can be concatenated into a vector.
The request cannot be done with more than 50 species at a time.
Use lapply, or any sort of loop when dealing with multiple
species (see example).}

\item{collectionCode}{Character. Any collection available at speciesLink.
Example: ALCB, E, INPA, MOBOT_BR.  Accepts a vector of names}

\item{country}{Character. Any country name. No ASCII characters allowed.
Accepts a vector of names}

\item{stateProvince}{Character. Any state or province. No ASCII characters
allowed. Accepts a vector of names}

\item{county}{Character. Any municipality name. No ASCII characters allowed.
Accepts a vector of names}

\item{Coordinates}{Character. Specify if records should have coordinates.
Default is "no check" but it also accepts "Yes", No", "Original",
"Automatic", "Blocked" or "no check"}

\item{CoordinatesQuality}{Character. Any character in "Good" or "Bad" to
select specific type of coordinates (argument is now deprecated)}

\item{Scope}{Character. Major group to be required. If NULL
searches all groups. Any in "p" (plants), "a" (animals), "m"
(microorganisms), "f" (fossil) or "b" (broad scope)}

\item{Synonyms}{If species names should be checked for synonyms in
a specific dictionary. Set to "sp2000" for search in
Catálogo da Vida species2000, "flora2020" for Flora do Brasil
2020, "MycoBank" for MycoBank, "AlgaeBase" for AlgaeBase, "DSMZ"
for  DSMZ Prokaryotic Nomenclature Up-to-Date, "Moure" for
Catálogo de Abelhas Moure or "no synonyms".}

\item{Typus}{Logical. If TRUE select only typus. Defaults to FALSE.}

\item{Images}{Logical. If TRUE select only records with images.
Defaults to FALSE.}

\item{RedList}{Character. Any of the following threat categories:
"CR" (critically endangered), "PEX" (Probably extinct), "EN"
(endangered), "EW" (Extinct in nature), "EX" (Extinct), "RE"
(Extinct in Brazil) or "VU" (Vulnerable)}

\item{MaxRecords}{Numeric. Maximum number of records to be
required. Default is 200, maximum is 5000.}

\item{file.format}{Character. The file extension to be used for
saving ('csv' or 'rds'). Default to 'csv'.}

\item{compress}{Logical. Should the file be compressed? Default to
FALSE.}
}
\value{
A data.frame with the search result, which can also be
saved on disk
}
\description{
This function access the Web service API v1.0 of the
\href{https://specieslink.net/ws/1.0/}{speciesLink} network and returns
species occurrence data
}
\details{
The time necessary to download records from the
speciesLink API will depend on the number of records and species
in the query and from the speed of the internet connection.

The speciesLink API does not allow the download of more than 5000
records at a time. So, to download records from species with
large number of records (>5000), you will probably need to make
the queries in a loop (see Examples).

Also, the speciesLink API does not allow the download of ~50 or
more taxa at a time. So to download records from larger lists of
species, you will also need to make the queries in a loop
(see Examples).
}
\examples{
\dontrun{
# Example for a single species, saving into a file called "ex01"

key <- (ADD HERE YOUR PERSONAL API KEY)

ex01 <- rspeciesLink(species =  c("Eugenia platyphylla"),
                     Scope = "p",
                     key = key)


# Use lapply or similar for more than 50 species

# Making a request for multiple species
sp <- c("Eugenia neoriedeliana", "Eugenia platyphylla")
sp_list <- lapply(sp, function(x) rspeciesLink(species = x,
                                              Scope = "p",
                                              basisOfRecord = "PreservedSpecimen",
                                              Synonyms = "flora2020",
                                              key = key))
# Adding species names to each element of the list
names(sp_list) = sp
# Binding all searchs and keeping a column w/ the named used in the search
sp_df <- bind_rows(sp_list, .id = "original_search")
}

}
\author{
Sara Mortara, João Vieira and Renato A. Ferreira de Lima
}
