% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strLoc.R
\name{strLoc}
\alias{strLoc}
\title{Format Locality}
\usage{
strLoc(x)
}
\arguments{
\item{x}{locality string}
}
\description{
Create the locality string used to search missing coordinates and to validade original coordinates.
}
