% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryData.R
\name{summaryData}
\alias{summaryData}
\title{Summary of Species Occurrence Data}
\usage{
summaryData(x, top = 5)
}
\arguments{
\item{x}{a data frame with the occurrence data.}

\item{top}{numerical. Number of groups (e.g. collections, families) to be
presented. Default to 5.}
}
\description{
This function provides a general summary of the occurrence data,
  including the number of records, biological collections, taxa and
  countries.
}
\details{
The summary output depends on the presence of some key columns in
  the input data frame \code{x}.

  The function prints the main summary tables related to the occurrence data.
  However, the list of tables that generate those summaries can be saved into an
  object, if needed.
}
