% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFlags.R
\name{summaryFlags}
\alias{summaryFlags}
\title{Summary Flags of Data Validation}
\usage{
summaryFlags(x)
}
\arguments{
\item{x}{a data frame with the occurrence data and the output of the
__plantR__ validation functions.}
}
\description{
This function provides the summary of the validatons steps of
  the species occurrence data, including localities, geographical
  coordinates, taxonomic confidence level and other outputs from the
  __plantR__ validation workflow.
}
\details{
The summary output depends on the presence of some key columns in
  the input data frame \code{x}.

  The function prints the main summary tables related to __plantR__
  validation workflow. However, the list of tables that generate those
  summaries can be saved into an object, if needed.
}
