% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdwgName.R
\name{tdwgName}
\alias{tdwgName}
\title{Format Name To TDWG Standard}
\usage{
tdwgName(name)
}
\arguments{
\item{name}{the character string.}
}
\value{
the character string \code{name} in the TDWG format. If only one name
is given, the function returns \code{name} with the first letter capitalized.
}
\description{
Put the the collector or determiner name in the TDWG format.
}
\details{
The function puts the name of a person into the format suggested by
the TDWG <International Working Group on Taxonomic Databases for Plant Sciences>.
The standard notation is: last name, followed by a comma and then the initials,
separated by points (e.g. Hatschbach, G.G.). Currenlty, the function remove name
prefixs or prepositions (e.g. de, dos, van, ter, ...). The function is relatively
stable regarding the input format and spacing, but it may not work in all cases,
particularly if the string provided already contain commas.
}
\examples{
# Simple name
  tdwgName("Al Gentry")
  tdwgName("Gert Guenther Hatschbach")
  tdwgName("Gert G. Hatschbach")
  tdwgName("HATSCHBACH, G.G.")
  tdwgName("HATSCHBACH, G. G.")
  tdwgName("G. G. HATSCHBACH")
# Name with generational suffixes
  tdwgName("Hermogenes Leitao Filho")
  tdwgName("Leitao Filho, H.")
  tdwgName("Leitao Filho, H.F.")
  tdwgName("S.J. Filho Neto")
# Compound last name
  tdwgName("Augustin Saint-hilaire")
  tdwgName("Saint-Hilaire A.")
# Unusual formatting (function won't always work)
  tdwgName("Cesar Sandro, Esteves, F") # one author, two commas: fails to get the right last name
  tdwgName("Mendonca Filho, C.V. Neto, F.C.C.") # two authors not properly separated: combine names of both authors
  tdwgName("Cyl Farney Catarino de Sa") # small last name, no comma: output correct
  tdwgName("Sa, Cyl") # small last name, no comma: output incorrect (inverted)
}
\references{
Willemse, L.P., van Welzen, P.C. & Mols, J.B. (2008). Standardisation in data-entry across databases: Avoiding Babylonian confusion. Taxon 57(2): 343-345.
}
\author{
Lima, R.A.F. & ter Steege, H.
}
