% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateCoord.R
\name{validateCoord}
\alias{validateCoord}
\title{Validate Geographical Coordinates}
\usage{
validateCoord(
  x = occs,
  lon = "decimalLongitude.new",
  lat = "decimalLatitude.new"
)
}
\arguments{
\item{x}{Data.frame with coordinates in decimal degrees.}

\item{lon}{Column with the longitude to be validated}

\item{lat}{Column with the latitude to be validated}
}
\description{
This function
}
\author{
Andrea Sánchez-Tapia, Sara Mortara & Renato A. F. de Lima
}
