% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateLoc.R
\name{validateLoc}
\alias{validateLoc}
\title{Validate Occurrence Locality Information}
\usage{
validateLoc(x)
}
\arguments{
\item{x}{Data frame.}
}
\description{
Compare the resolution of the original locality information provided with the
one retrieved from the locality gazetteer.
}
