% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workCoord.R
\name{workCoord}
\alias{workCoord}
\title{Get Working Geographical Coordinates}
\usage{
workCoord(x)
}
\arguments{
\item{x}{a data.frame containing the original coordinates and the coordinates
obtained from a gazetteer.}
}
\value{
The data frame \code{x} with the columns 'decimalLatitude.new' and
  'decimalLongitude.new', which contain the geographical coordinates of the
  gazetteer at the best resolution (i.e. administrative level) available.
}
\description{
This function assess the availability and resolution of
  the original geographical coordinates of the occurrence and it replaces
  missing coordinates by the coordinates obtained from the gazetteer.
}
\details{
The function was initially designed as part of a larger routine to
  edit and validate geographical coordinates of plant occurrence data. It is
  possible to use it separately, but it may be easier to use it under the
  workflow presented in `plantR` manual. If used separately, users must
  provide a data frame with at least six columns ('decimalLatitude',
  'decimalLongitude', 'latitude.gazetteer', 'longitude.gazetteer',
  'resolution.gazetteer', 'decimalLatitude.new' and 'decimalLongitude.new).
  See the examples below.

  (Explain how the resolution of the original coordinate is obtained)
}
\examples{

## To be included ##

(coords <- data.frame(decimalLatitude = c(-23.475389, NA),
                     decimalLongitude = c(-47.123768, NA),
                     latitude.gazetteer = c(-23.524052, -23.524052),
                     longitude.gazetteer = c(-47.122207, -47.122207),
                     resolution.gazetteer = c("county","county"),
                     decimalLatitude.new = c(-23.475389, NA),
                     decimalLongitude.new = c(-47.123768, NA)
                     ))

workCoord(coords)

}
\author{
Renato A. F. de Lima
}
