% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{worldMap}
\alias{worldMap}
\title{World Map}
\format{
A Multipolygon 'sf' object with 256 features
}
\source{
\url{https://gadm.org}
}
\usage{
data(worldMap)
}
\description{
Map used to perform the validation of the original geographical coordinates.
Country names were edited to match the \strong{plantR} default gazetteer and
original polygons were simplified (tolerance 0.001 decimal degrees).
Reference system: WSG84
}
\keyword{datasets}
